refactor(core): split the 4k-line navigation.cpp into product TUs
- navigation_internal.hpp: former anonymous-namespace helpers (SSR/CLAS
  base-clock hygiene, epoch parsers, IONEX interpolation, broadcast
  orbit math) as inline functions in libgnss::navigation_internal
- ssr_products.cpp: SSRProducts correction ingest / interpolation /
  held-token accessors (~1.9k lines)
- precise_products.cpp: PreciseProducts SP3/clock loading and Neville
  interpolation
- ionex_dcb_products.cpp: IONEXProducts and DCBProducts loaders
- navigation.cpp keeps NavigationData/Ephemeris and the ionosphere/
  troposphere delay models

Verified: ctest 71/71, byte-identical rtk_positioning output on PPC
tokyo/run1 vs pre-split baseline, MADOCALIB parity suite 17/17.
