+++ src/core/navigation.cpp
@@ -12,1018 +12,11 @@
 #include <utility>
 #include <vector>
 
+#include "navigation_internal.hpp"
+
 namespace libgnss {
 
-namespace {
-
-constexpr double kGpsWeekSeconds = 604800.0;
-constexpr double kHalfGpsWeekSeconds = 302400.0;
-constexpr double kEarthRotationRate = 7.2921151467e-5;
-constexpr double kEarthMu = 3.986005e14;
-constexpr double kGalileoEarthMu = 3.986004418e14;
-constexpr double kGlonassEarthMu = 3.9860044e14;
-constexpr double kGlonassJ2 = 1.0826257e-3;
-constexpr double kGlonassEarthRadius = 6378136.0;
-constexpr double kGlonassEarthRotationRate = 7.292115e-5;
-constexpr double kGlonassIntegrationStep = 60.0;
-constexpr double kBeiDouEarthRotationRate = 7.292115e-5;
-constexpr double kBeiDouEarthMu = 3.986004418e14;
-constexpr double kRelativityF = -4.442807633e-10;
-constexpr double kWgs84A = 6378137.0;
-constexpr double kWgs84E2 = 6.69437999014e-3;
-constexpr double kBeiDouGeoSin5Deg = -0.0871557427476582;
-constexpr double kBeiDouGeoCos5Deg = 0.9961946980917456;
-constexpr double kPreciseInterpolationGapSeconds = 900.0;
-// CLASLIB STECVALIDAGE (cssr.h:79) — compact regional STEC held across this window.
-constexpr double kClasStecValidAgeSeconds = 3600.0;
-
-bool isQzssDedicatedBiasBankEntry(const SSROrbitClockCorrection& entry, bool phase) {
-    const bool valid =
-        phase ? (entry.phase_bias_valid && !entry.phase_bias_m.empty())
-              : (entry.code_bias_valid && !entry.code_bias_m.empty());
-    if (!valid) {
-        return false;
-    }
-    if (phase) {
-        if (entry.bias_network_id <= 0) {
-            return false;
-        }
-        for (const auto& [signal_id, bias_m] : entry.phase_bias_m) {
-            if (std::abs(bias_m) > 1e-12) {
-                return true;
-            }
-        }
-        return false;
-    }
-    return entry.orbit_valid || entry.clock_valid;
-}
-
-const SSROrbitClockCorrection* findQzssHeldPhaseBiasForServiceNetwork(
-    const std::vector<SSROrbitClockCorrection>& entries,
-    const GNSSTime& time,
-    int service_network_id,
-    size_t scan_end_index) {
-    if (service_network_id <= 0 || scan_end_index == 0) {
-        return nullptr;
-    }
-    size_t group_end = scan_end_index;
-    while (group_end > 0) {
-        const size_t group_last = group_end - 1;
-        const GNSSTime group_time = entries[group_last].time;
-        if (time - group_time > kClasStecValidAgeSeconds + 1e-9) {
-            break;
-        }
-        size_t group_begin = group_last;
-        while (group_begin > 0 && entries[group_begin - 1].time == group_time) {
-            --group_begin;
-        }
-        for (size_t index = group_begin; index < group_end; ++index) {
-            const SSROrbitClockCorrection& entry = entries[index];
-            if (!isQzssDedicatedBiasBankEntry(entry, true)) {
-                continue;
-            }
-            if (entry.bias_network_id == service_network_id) {
-                return &entry;
-            }
-        }
-        group_end = group_begin;
-    }
-    return nullptr;
-}
-
-std::string trimCopy(const std::string& value) {
-    const auto begin = value.find_first_not_of(" \t\r\n");
-    if (begin == std::string::npos) {
-        return "";
-    }
-    const auto end = value.find_last_not_of(" \t\r\n");
-    return value.substr(begin, end - begin + 1);
-}
-
-GNSSSystem systemFromChar(char system_char) {
-    switch (system_char) {
-        case 'G': return GNSSSystem::GPS;
-        case 'R': return GNSSSystem::GLONASS;
-        case 'E': return GNSSSystem::Galileo;
-        case 'C': return GNSSSystem::BeiDou;
-        case 'J': return GNSSSystem::QZSS;
-        case 'S': return GNSSSystem::SBAS;
-        case 'I': return GNSSSystem::NavIC;
-        default: return GNSSSystem::UNKNOWN;
-    }
-}
-
-bool parseSatelliteToken(const std::string& token, SatelliteId& satellite) {
-    if (token.size() < 2) {
-        return false;
-    }
-    const GNSSSystem system = systemFromChar(token[0]);
-    if (system == GNSSSystem::UNKNOWN) {
-        return false;
-    }
-    try {
-        satellite = SatelliteId(system, static_cast<uint8_t>(std::stoi(token.substr(1))));
-        return true;
-    } catch (const std::exception&) {
-        return false;
-    }
-}
-
-bool parseCompactSsrSatelliteToken(const std::string& token, SatelliteId& satellite) {
-    if (pppEnvOverrides().clas_qzss_s_prn_fix &&
-        token.size() >= 2 &&
-        token[0] == 'S') {
-        try {
-            const int prn = std::stoi(token.substr(1));
-            if (prn >= 120 && prn <= 122) {
-                satellite = SatelliteId(
-                    GNSSSystem::QZSS,
-                    static_cast<uint8_t>(1 + (prn - 120)));
-                return true;
-            }
-        } catch (const std::exception&) {
-        }
-    }
-    return parseSatelliteToken(token, satellite);
-}
-
-std::string remapCompactSsrAtmosKey(std::string key) {
-    if (!pppEnvOverrides().clas_qzss_s_prn_fix) {
-        return key;
-    }
-    const std::pair<const char*, const char*> remaps[] = {
-        {":S120", ":J01"},
-        {":S121", ":J02"},
-        {":S122", ":J03"},
-    };
-    for (const auto& [from, to] : remaps) {
-        size_t pos = 0;
-        while ((pos = key.find(from, pos)) != std::string::npos) {
-            key.replace(pos, std::string(from).size(), to);
-            pos += std::string(to).size();
-        }
-    }
-    return key;
-}
-
-int parsePositiveIntToken(const std::map<std::string, std::string>& tokens,
-                          const std::string& key) {
-    const auto it = tokens.find(key);
-    if (it == tokens.end()) {
-        return 0;
-    }
-    try {
-        const int value = std::stoi(it->second);
-        return value > 0 ? value : 0;
-    } catch (const std::exception&) {
-        return 0;
-    }
-}
-
-bool sameCorrectionVariant(const SSROrbitClockCorrection& lhs,
-                           const SSROrbitClockCorrection& rhs) {
-    return lhs.atmos_network_id == rhs.atmos_network_id &&
-           lhs.bias_network_id == rhs.bias_network_id;
-}
-
-bool clasQzssBaseClockHygieneActive(const SatelliteId& sat);
-bool atmosEmbeddedClockRow(const SSROrbitClockCorrection& entry);
-void promoteClaslibBaseClockFields(SSROrbitClockCorrection& correction);
-
-void mergeCorrectionFields(SSROrbitClockCorrection& target,
-                           const SSROrbitClockCorrection& correction) {
-    if (correction.clock_valid && correction.clock_network_id == 0 &&
-        !target.mrtklib_base_clock_valid) {
-        target.mrtklib_base_clock_correction_m = correction.clock_correction_m;
-        target.mrtklib_base_clock_reference_time =
-            correction.clock_reference_time.week != 0
-                ? correction.clock_reference_time
-                : correction.time;
-        target.mrtklib_base_clock_valid = true;
-    }
-    if (correction.orbit_valid) {
-        target.orbit_correction_ecef = correction.orbit_correction_ecef;
-        target.orbit_valid = true;
-        target.orbit_reference_time = correction.orbit_reference_time.week != 0 ?
-            correction.orbit_reference_time : correction.time;
-        target.iode = correction.iode;
-        target.ssr_orbit_iod = correction.ssr_orbit_iod;
-    }
-    if (correction.clock_valid) {
-        if (correction.clock_network_id == 0) {
-            target.clock_correction_m = correction.clock_correction_m;
-            const bool skip_atmos_base =
-                clasQzssBaseClockHygieneActive(target.satellite) &&
-                atmosEmbeddedClockRow(correction);
-            if (!skip_atmos_base) {
-                target.base_clock_correction_m = correction.clock_correction_m;
-                target.base_clock_valid = true;
-                target.base_clock_reference_time = correction.clock_reference_time.week != 0 ?
-                    correction.clock_reference_time : correction.time;
-            }
-        } else {
-            target.clock_correction_m = correction.clock_correction_m;
-            if (clasQzssBaseClockHygieneActive(target.satellite) &&
-                !atmosEmbeddedClockRow(correction)) {
-                target.base_clock_correction_m = correction.clock_correction_m;
-                target.base_clock_valid = true;
-                target.base_clock_reference_time = correction.clock_reference_time.week != 0 ?
-                    correction.clock_reference_time : correction.time;
-            }
-        }
-        target.clock_valid = true;
-        target.clock_reference_time = correction.clock_reference_time.week != 0 ?
-            correction.clock_reference_time : correction.time;
-        target.ssr_clock_iod = correction.ssr_clock_iod;
-    }
-    if (correction.clock_withdrawn) {
-        target.clock_correction_m = 0.0;
-        target.clock_valid = false;
-        target.base_clock_correction_m = 0.0;
-        target.base_clock_valid = false;
-        target.mrtklib_base_clock_correction_m = 0.0;
-        target.mrtklib_base_clock_valid = false;
-        target.clock_withdrawn = true;
-        target.clock_reference_time = correction.time;
-    }
-    if (correction.ura_valid) {
-        target.ura_sigma_m = correction.ura_sigma_m;
-        target.ura_valid = true;
-    }
-    if (correction.code_bias_valid) {
-        for (const auto& [signal_id, bias_m] : correction.code_bias_m) {
-            target.code_bias_m[signal_id] = bias_m;
-        }
-        target.code_bias_valid = !target.code_bias_m.empty();
-        target.code_bias_rtklib_m = correction.code_bias_rtklib_m;
-    }
-    if (correction.phase_bias_valid) {
-        for (const auto& [signal_id, bias_m] : correction.phase_bias_m) {
-            target.phase_bias_m[signal_id] = bias_m;
-        }
-        for (const auto& [signal_id, discnt] : correction.phase_bias_discnt) {
-            target.phase_bias_discnt[signal_id] = discnt;
-        }
-        target.phase_bias_valid = !target.phase_bias_m.empty();
-        target.phase_bias_rtklib_m = correction.phase_bias_rtklib_m;
-    }
-    if (correction.atmos_valid) {
-        for (const auto& [token, value] : correction.atmos_tokens) {
-            target.atmos_tokens[token] = value;
-        }
-        target.atmos_valid = !target.atmos_tokens.empty();
-    }
-}
-
-GNSSTime orbitReferenceTime(const SSROrbitClockCorrection& correction) {
-    return correction.orbit_reference_time.week != 0 ?
-        correction.orbit_reference_time : correction.time;
-}
-
-GNSSTime clockReferenceTime(const SSROrbitClockCorrection& correction) {
-    return correction.clock_reference_time.week != 0 ?
-        correction.clock_reference_time : correction.time;
-}
-
-double sampleBaseClockM(const SSROrbitClockCorrection& sample, bool& valid) {
-    if (sample.base_clock_valid) {
-        valid = true;
-        return sample.base_clock_correction_m;
-    }
-    valid = sample.clock_valid;
-    return sample.clock_correction_m;
-}
-
-void assignBaseClockOutputs(const SSROrbitClockCorrection& sample,
-                            double merged_clock_m,
-                            double* base_clock_correction_m,
-                            bool* base_clock_valid) {
-    if (base_clock_correction_m == nullptr || base_clock_valid == nullptr) {
-        return;
-    }
-    bool valid = false;
-    const double base_m = sampleBaseClockM(sample, valid);
-    if (valid) {
-        *base_clock_correction_m = base_m;
-        *base_clock_valid = true;
-        return;
-    }
-    *base_clock_correction_m = merged_clock_m;
-    *base_clock_valid = false;
-}
-
-void interpolateBaseClockOutputs(const SSROrbitClockCorrection* before,
-                                 const SSROrbitClockCorrection* after,
-                                 double alpha,
-                                 double merged_clock_m,
-                                 double* base_clock_correction_m,
-                                 bool* base_clock_valid,
-                                 SSRClockSelectionPolicy clock_selection_policy =
-                                     SSRClockSelectionPolicy::MergedInterpolate) {
-    if (base_clock_correction_m == nullptr || base_clock_valid == nullptr) {
-        return;
-    }
-    if (clock_selection_policy == SSRClockSelectionPolicy::ClaslibBaseHold ||
-        clock_selection_policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-        const SSROrbitClockCorrection* sample = before != nullptr ? before : after;
-        if (sample != nullptr) {
-            assignBaseClockOutputs(*sample, merged_clock_m,
-                                   base_clock_correction_m, base_clock_valid);
-        } else {
-            *base_clock_correction_m = merged_clock_m;
-            *base_clock_valid = false;
-        }
-        return;
-    }
-    const bool before_base = before != nullptr && before->base_clock_valid;
-    const bool after_base = after != nullptr && after->base_clock_valid;
-    if (before_base && after_base) {
-        *base_clock_correction_m =
-            before->base_clock_correction_m +
-            alpha * (after->base_clock_correction_m - before->base_clock_correction_m);
-        *base_clock_valid = true;
-        return;
-    }
-    if (before_base) {
-        *base_clock_correction_m = before->base_clock_correction_m;
-        *base_clock_valid = true;
-        return;
-    }
-    if (after_base) {
-        *base_clock_correction_m = after->base_clock_correction_m;
-        *base_clock_valid = true;
-        return;
-    }
-    *base_clock_correction_m = merged_clock_m;
-    *base_clock_valid = false;
-}
-
-constexpr double kClaslibMaxOrbitAgeFromObsSeconds = 180.0;
-constexpr double kClaslibMaxClockAgeFromObsSeconds = 30.0;
-constexpr double kClaslibMaxClockAgeFromOrbitSeconds = 30.0;
-// The compact bias epoch is the correction reference epoch, not its L6
-// reception epoch. In the MRTKLIB stream the ST6 bank becomes causal at the
-// half-cycle point (reference TOW + 15 s); earlier observations still use the
-// preceding 30 s bank.
-constexpr double kMrtklibBiasReceptionLagSeconds = 15.0;
-
-bool clasQzssBaseClockHygieneActive(const SatelliteId& sat) {
-    return pppEnvOverrides().clas_base_clock_parity &&
-           pppEnvOverrides().clas_qzss_s_prn_fix &&
-           sat.system == GNSSSystem::QZSS && sat.prn >= 1 && sat.prn <= 3;
-}
-
-bool atmosEmbeddedClockRow(const SSROrbitClockCorrection& entry) {
-    // Network-bank rows (clock_network_id>0) may carry forward-filled atmos
-    // tokens from CSV ingest; only reject atmos-primary embedded dclock rows.
-    return entry.atmos_valid && !entry.atmos_tokens.empty() &&
-           entry.clock_network_id == 0;
-}
-
-bool claslibBaseClockCandidate(const SatelliteId& sat,
-                               const SSROrbitClockCorrection& entry) {
-    if (clasQzssBaseClockHygieneActive(sat) && atmosEmbeddedClockRow(entry)) {
-        return false;
-    }
-    if (entry.base_clock_valid) {
-        return true;
-    }
-    if (clasQzssBaseClockHygieneActive(sat) && entry.clock_valid &&
-        entry.clock_network_id > 0 && !atmosEmbeddedClockRow(entry)) {
-        return true;
-    }
-    return false;
-}
-
-double claslibBaseClockCorrectionM(const SSROrbitClockCorrection& entry) {
-    if (entry.clock_valid && entry.clock_network_id > 0) {
-        return entry.clock_correction_m;
-    }
-    return entry.base_clock_correction_m;
-}
-
-void promoteClaslibBaseClockFields(SSROrbitClockCorrection& correction) {
-    if (!correction.clock_valid) {
-        return;
-    }
-    const bool hygiene = clasQzssBaseClockHygieneActive(correction.satellite);
-    if (hygiene && atmosEmbeddedClockRow(correction)) {
-        return;
-    }
-    if (correction.clock_network_id == 0) {
-        correction.base_clock_correction_m = correction.clock_correction_m;
-        correction.base_clock_valid = true;
-        correction.base_clock_reference_time =
-            correction.clock_reference_time.week != 0 ?
-                correction.clock_reference_time : correction.time;
-        return;
-    }
-    if (hygiene && correction.clock_network_id > 0 && !atmosEmbeddedClockRow(correction)) {
-        correction.base_clock_correction_m = correction.clock_correction_m;
-        correction.base_clock_valid = true;
-        correction.base_clock_reference_time =
-            correction.clock_reference_time.week != 0 ?
-                correction.clock_reference_time : correction.time;
-    }
-}
-
-struct ClaslibClockPick {
-    bool valid = false;
-    bool withdrawn = false;
-    double clock_correction_m = 0.0;
-    GNSSTime clock_time;
-    GNSSTime clock_reference_time;
-    int ssr_clock_iod = -1;
-};
-
-ClaslibClockPick pickClaslibBaseClock(
-    const SatelliteId& satellite,
-    const std::vector<SSROrbitClockCorrection>& entries,
-    const GNSSTime& obs_time,
-    SSRClockSelectionPolicy policy) {
-    const SSROrbitClockCorrection* orbit_ref = nullptr;
-    GNSSTime orbit_ref_time;
-    for (const auto& entry : entries) {
-        if (!entry.orbit_valid) {
-            continue;
-        }
-        const GNSSTime orbit_time = orbitReferenceTime(entry);
-        const double obs_age = obs_time - orbit_time;
-        if (obs_age < 0.0 || obs_age > kClaslibMaxOrbitAgeFromObsSeconds) {
-            continue;
-        }
-        if (orbit_ref == nullptr || orbit_time > orbit_ref_time) {
-            orbit_ref = &entry;
-            orbit_ref_time = orbit_time;
-        }
-    }
-    if (orbit_ref == nullptr) {
-        return {};
-    }
-
-    const bool galileo_clock_withdrawal =
-        policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold &&
-        satellite.system == GNSSSystem::Galileo;
-    const SSROrbitClockCorrection* best = nullptr;
-    auto clock_end = std::upper_bound(
-        entries.begin(),
-        entries.end(),
-        obs_time,
-        [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-            return lhs < rhs.time;
-        });
-    while (clock_end != entries.begin()) {
-        const SSROrbitClockCorrection& entry = *--clock_end;
-        const double obs_age = obs_time - entry.time;
-        if (obs_age > kClaslibMaxClockAgeFromObsSeconds) {
-            break;
-        }
-        if (best != nullptr && entry.time < best->time) {
-            break;
-        }
-        const bool literal =
-            policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold;
-        const bool candidate_valid = literal
-            ? entry.mrtklib_base_clock_valid
-            : claslibBaseClockCandidate(satellite, entry);
-        if (!candidate_valid &&
-            !(galileo_clock_withdrawal && entry.clock_withdrawn)) {
-            continue;
-        }
-
-        const GNSSTime clock_time = entry.time;
-        if (clock_time > obs_time) {
-            continue;
-        }
-        if (obs_age < 0.0 || obs_age > kClaslibMaxClockAgeFromObsSeconds) {
-            continue;
-        }
-        const double orbit_age = clock_time - orbit_ref_time;
-        if (orbit_age < 0.0 || orbit_age >= kClaslibMaxClockAgeFromOrbitSeconds) {
-            continue;
-        }
-        if (best == nullptr || clock_time > best->time ||
-            (clock_time == best->time && galileo_clock_withdrawal &&
-             entry.clock_withdrawn && !best->clock_withdrawn) ||
-            (clock_time == best->time &&
-             clasQzssBaseClockHygieneActive(satellite) &&
-             entry.clock_network_id > 0 && best->clock_network_id == 0)) {
-            best = &entry;
-        }
-    }
-    if (best == nullptr) {
-        return {};
-    }
-
-    if (galileo_clock_withdrawal && best->clock_withdrawn) {
-        ClaslibClockPick pick;
-        pick.withdrawn = true;
-        pick.clock_time = best->time;
-        pick.clock_reference_time = best->time;
-        return pick;
-    }
-
-    ClaslibClockPick pick;
-    pick.valid = true;
-    const bool literal =
-        policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold;
-    pick.clock_correction_m = literal
-        ? best->mrtklib_base_clock_correction_m
-        : claslibBaseClockCorrectionM(*best);
-    pick.clock_time = best->time;
-    pick.clock_reference_time = literal
-        ? best->mrtklib_base_clock_reference_time
-        : clockReferenceTime(*best);
-    pick.ssr_clock_iod = best->ssr_clock_iod;
-    return pick;
-}
-
-bool applyClaslibClockSelection(
-    SSRClockSelectionPolicy clock_selection_policy,
-    const SatelliteId& satellite,
-    const std::vector<SSROrbitClockCorrection>& entries,
-    const GNSSTime& time,
-    double& clock_correction_m,
-    double* base_clock_correction_m,
-    bool* base_clock_valid,
-    GNSSTime* clock_reference_time,
-    SSRCorrectionStatus* status,
-    bool base_result) {
-    if (clock_selection_policy != SSRClockSelectionPolicy::ClaslibBaseHold &&
-        clock_selection_policy != SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-        return base_result;
-    }
-
-    const ClaslibClockPick pick =
-        pickClaslibBaseClock(satellite, entries, time, clock_selection_policy);
-    if (!pick.valid) {
-        if (clasQzssBaseClockHygieneActive(satellite) && base_result &&
-            (status == nullptr || status->orbit_valid)) {
-            return true;
-        }
-        clock_correction_m = 0.0;
-        if (base_clock_correction_m != nullptr && base_clock_valid != nullptr) {
-            *base_clock_correction_m = 0.0;
-            *base_clock_valid = false;
-        }
-        if (clock_reference_time != nullptr) {
-            *clock_reference_time = GNSSTime();
-        }
-        if (status != nullptr) {
-            status->clock_valid = false;
-            status->clock_withdrawn = pick.withdrawn;
-            if (pick.withdrawn) {
-                status->clock_reference_time = pick.clock_reference_time;
-            }
-        }
-        return false;
-    }
-
-    clock_correction_m = pick.clock_correction_m;
-    if (base_clock_correction_m != nullptr && base_clock_valid != nullptr) {
-        *base_clock_correction_m = pick.clock_correction_m;
-        *base_clock_valid = true;
-    }
-    if (clock_reference_time != nullptr) {
-        *clock_reference_time = pick.clock_time;
-    }
-    if (status != nullptr) {
-        status->clock_valid = true;
-        status->clock_reference_time = pick.clock_reference_time;
-        status->ssr_clock_iod = pick.ssr_clock_iod;
-    }
-    const bool orbit_ok = status == nullptr || status->orbit_valid;
-    return pick.valid && orbit_ok;
-}
-
-bool parseEpochFields(int year,
-                      int month,
-                      int day,
-                      int hour,
-                      int minute,
-                      double second,
-                      GNSSTime& time) {
-    std::tm epoch_tm{};
-    epoch_tm.tm_year = year - 1900;
-    epoch_tm.tm_mon = month - 1;
-    epoch_tm.tm_mday = day;
-    epoch_tm.tm_hour = hour;
-    epoch_tm.tm_min = minute;
-    epoch_tm.tm_sec = static_cast<int>(std::floor(second));
-#if defined(_WIN32)
-    const time_t seconds_since_unix = _mkgmtime(&epoch_tm);
-#else
-    const time_t seconds_since_unix = timegm(&epoch_tm);
-#endif
-    if (seconds_since_unix == static_cast<time_t>(-1)) {
-        return false;
-    }
-    const auto tp = std::chrono::system_clock::from_time_t(seconds_since_unix) +
-        std::chrono::microseconds(
-            static_cast<long long>(std::llround((second - std::floor(second)) * 1e6)));
-    time = GNSSTime::fromSystemTime(tp);
-    return true;
-}
-
-bool parseSp3EpochLine(const std::string& line, GNSSTime& time) {
-    if (line.empty() || line[0] != '*') {
-        return false;
-    }
-    std::istringstream stream(line.substr(1));
-    int year = 0;
-    int month = 0;
-    int day = 0;
-    int hour = 0;
-    int minute = 0;
-    double second = 0.0;
-    if (!(stream >> year >> month >> day >> hour >> minute >> second)) {
-        return false;
-    }
-    return parseEpochFields(year, month, day, hour, minute, second, time);
-}
-
-bool parseClockEpochTokens(std::istringstream& stream, GNSSTime& time) {
-    int year = 0;
-    int month = 0;
-    int day = 0;
-    int hour = 0;
-    int minute = 0;
-    double second = 0.0;
-    if (!(stream >> year >> month >> day >> hour >> minute >> second)) {
-        return false;
-    }
-    return parseEpochFields(year, month, day, hour, minute, second, time);
-}
-
-bool parseIonexEpochLine(const std::string& line, GNSSTime& time) {
-    const std::string epoch_text = trimCopy(line.substr(0, std::min<size_t>(43U, line.size())));
-    if (epoch_text.empty()) {
-        return false;
-    }
-    std::istringstream stream(epoch_text);
-    int year = 0;
-    int month = 0;
-    int day = 0;
-    int hour = 0;
-    int minute = 0;
-    double second = 0.0;
-    if (!(stream >> year >> month >> day >> hour >> minute >> second)) {
-        return false;
-    }
-    return parseEpochFields(year, month, day, hour, minute, second, time);
-}
-
-bool parseFixedFieldDoubles(const std::string& line,
-                            size_t end_column,
-                            std::vector<double>& values) {
-    values.clear();
-    std::istringstream stream(line.substr(0, std::min(end_column, line.size())));
-    double value = 0.0;
-    while (stream >> value) {
-        values.push_back(value);
-    }
-    return !values.empty();
-}
-
-double normalizeLongitudeToGrid(double lon_deg, double lon_start_deg, double lon_end_deg) {
-    double normalized = lon_deg;
-    const double min_lon = std::min(lon_start_deg, lon_end_deg);
-    const double max_lon = std::max(lon_start_deg, lon_end_deg);
-    while (normalized < min_lon - 1e-9) {
-        normalized += 360.0;
-    }
-    while (normalized > max_lon + 1e-9) {
-        normalized -= 360.0;
-    }
-    return normalized;
-}
-
-bool interpolateIonexRow(const IONEXLatitudeRow& row, double longitude_deg, double& tecu) {
-    if (row.values_tecu.empty() || std::abs(row.longitude_step_deg) < 1e-12) {
-        return false;
-    }
-    const double normalized_lon =
-        normalizeLongitudeToGrid(longitude_deg, row.longitude_start_deg, row.longitude_end_deg);
-    const double index = (normalized_lon - row.longitude_start_deg) / row.longitude_step_deg;
-    if (index < -1e-9 || index > static_cast<double>(row.values_tecu.size() - 1U) + 1e-9) {
-        return false;
-    }
-
-    const double clamped = std::max(0.0, std::min(index, static_cast<double>(row.values_tecu.size() - 1U)));
-    const size_t left = static_cast<size_t>(std::floor(clamped));
-    const size_t right = std::min(left + 1U, row.values_tecu.size() - 1U);
-    const double alpha = clamped - static_cast<double>(left);
-    const double left_value = row.values_tecu[left];
-    const double right_value = row.values_tecu[right];
-    if (!std::isfinite(left_value) || !std::isfinite(right_value)) {
-        return false;
-    }
-    tecu = left_value + alpha * (right_value - left_value);
-    return true;
-}
-
-bool interpolateIonexMap(const IONEXMap& map,
-                         double latitude_deg,
-                         double longitude_deg,
-                         double& tecu) {
-    if (map.rows.empty()) {
-        return false;
-    }
-
-    const IONEXLatitudeRow* before = nullptr;
-    const IONEXLatitudeRow* after = nullptr;
-    for (const auto& row : map.rows) {
-        if (row.latitude_deg <= latitude_deg) {
-            if (before == nullptr || row.latitude_deg > before->latitude_deg) {
-                before = &row;
-            }
-        }
-        if (row.latitude_deg >= latitude_deg) {
-            if (after == nullptr || row.latitude_deg < after->latitude_deg) {
-                after = &row;
-            }
-        }
-    }
-    if (before == nullptr) {
-        before = &map.rows.front();
-    }
-    if (after == nullptr) {
-        after = &map.rows.back();
-    }
-
-    double before_tecu = 0.0;
-    double after_tecu = 0.0;
-    if (before == after) {
-        return interpolateIonexRow(*before, longitude_deg, tecu);
-    }
-    if (!interpolateIonexRow(*before, longitude_deg, before_tecu) ||
-        !interpolateIonexRow(*after, longitude_deg, after_tecu)) {
-        return false;
-    }
-
-    const double lat_delta = after->latitude_deg - before->latitude_deg;
-    if (std::abs(lat_delta) < 1e-12) {
-        tecu = before_tecu;
-        return true;
-    }
-    const double alpha = (latitude_deg - before->latitude_deg) / lat_delta;
-    tecu = before_tecu + alpha * (after_tecu - before_tecu);
-    return true;
-}
-
-template <typename EntryType>
-const EntryType* findEntryAtOrBefore(const std::vector<EntryType>& entries, const GNSSTime& time) {
-    auto upper = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        time,
-        [](const EntryType& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-    if (upper == entries.end()) {
-        return entries.empty() ? nullptr : &entries.back();
-    }
-    if (upper->time == time) {
-        return &(*upper);
-    }
-    if (upper == entries.begin()) {
-        return nullptr;
-    }
-    return &(*(upper - 1));
-}
-
-template <typename EntryType>
-const EntryType* findEntryAtOrAfter(const std::vector<EntryType>& entries, const GNSSTime& time) {
-    auto lower = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        time,
-        [](const EntryType& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-    if (lower == entries.end()) {
-        return nullptr;
-    }
-    return &(*lower);
-}
-
-double normalizeGpsTime(double dt) {
-    while (dt > kHalfGpsWeekSeconds) {
-        dt -= kGpsWeekSeconds;
-    }
-    while (dt < -kHalfGpsWeekSeconds) {
-        dt += kGpsWeekSeconds;
-    }
-    return dt;
-}
-
-double solveKepler(double mean_anomaly, double eccentricity) {
-    double E = mean_anomaly;
-    for (int i = 0; i < 30; ++i) {  // RTKLIB: MAX_ITER_KEPLER=30
-        double Ek = E;
-        E -= (E - eccentricity * std::sin(E) - mean_anomaly) / (1.0 - eccentricity * std::cos(E));
-        if (std::abs(E - Ek) < 1e-14) break;  // RTKLIB: RTOL_KEPLER=1e-14
-    }
-    return E;
-}
-
-bool isBeiDouGeoPrn(uint8_t prn) {
-    return prn <= 5 || (prn >= 59 && prn <= 63);
-}
-
-bool computeBroadcastState(const Ephemeris& eph,
-                           const GNSSTime& time,
-                           Vector3d& pos,
-                           double& clock_bias,
-                           double& clock_drift,
-                           bool use_mrtklib_galileo_mu) {
-    if (!eph.valid || eph.sqrt_a <= 0.0) {
-        return false;
-    }
-
-    const double a = eph.sqrt_a * eph.sqrt_a;
-    const double tk = normalizeGpsTime(time - eph.toe);
-    const double tc = normalizeGpsTime(time - eph.toc);
-    const bool is_beidou = eph.satellite.system == GNSSSystem::BeiDou;
-    const bool is_galileo =
-        use_mrtklib_galileo_mu &&
-        eph.satellite.system == GNSSSystem::Galileo;
-    const double earth_mu = is_beidou ? kBeiDouEarthMu
-                                      : (is_galileo ? kGalileoEarthMu : kEarthMu);
-    const double earth_rotation = is_beidou ? kBeiDouEarthRotationRate : kEarthRotationRate;
-
-    const double n0 = std::sqrt(earth_mu / (a * a * a));
-    const double n = n0 + eph.delta_n;
-    const double mean_anomaly = eph.m0 + n * tk;
-    const double eccentric_anomaly = solveKepler(mean_anomaly, eph.e);
-
-    const double sin_e = std::sin(eccentric_anomaly);
-    const double cos_e = std::cos(eccentric_anomaly);
-    const double sqrt_one_minus_e2 = std::sqrt(1.0 - eph.e * eph.e);
-    const double true_anomaly = std::atan2(sqrt_one_minus_e2 * sin_e, cos_e - eph.e);
-    const double phi = true_anomaly + eph.omega;
-
-    const double two_phi = 2.0 * phi;
-    const double du = eph.cus * std::sin(two_phi) + eph.cuc * std::cos(two_phi);
-    const double dr = eph.crs * std::sin(two_phi) + eph.crc * std::cos(two_phi);
-    const double di = eph.cis * std::sin(two_phi) + eph.cic * std::cos(two_phi);
-
-    const double u = phi + du;
-    const double r = a * (1.0 - eph.e * cos_e) + dr;
-    const double inclination = eph.i0 + eph.idot * tk + di;
-    const double toe_seconds = eph.toes != 0.0 ? eph.toes : eph.toe.tow;
-
-    const double x_orb = r * std::cos(u);
-    const double y_orb = r * std::sin(u);
-
-    const double cos_i = std::cos(inclination);
-    const double sin_i = std::sin(inclination);
-
-    if (is_beidou && isBeiDouGeoPrn(eph.satellite.prn)) {
-        const double omega_geo =
-            eph.omega0 + eph.omega_dot * tk - earth_rotation * toe_seconds;
-        const double cos_omega = std::cos(omega_geo);
-        const double sin_omega = std::sin(omega_geo);
-        const double xg = x_orb * cos_omega - y_orb * cos_i * sin_omega;
-        const double yg = x_orb * sin_omega + y_orb * cos_i * cos_omega;
-        const double zg = y_orb * sin_i;
-        const double sin_rot = std::sin(earth_rotation * tk);
-        const double cos_rot = std::cos(earth_rotation * tk);
-        pos(0) = xg * cos_rot + yg * sin_rot * kBeiDouGeoCos5Deg +
-                 zg * sin_rot * kBeiDouGeoSin5Deg;
-        pos(1) = -xg * sin_rot + yg * cos_rot * kBeiDouGeoCos5Deg +
-                 zg * cos_rot * kBeiDouGeoSin5Deg;
-        pos(2) = -yg * kBeiDouGeoSin5Deg + zg * kBeiDouGeoCos5Deg;
-    } else {
-        const double omega = eph.omega0 + (eph.omega_dot - earth_rotation) * tk
-            - earth_rotation * toe_seconds;
-        const double cos_omega = std::cos(omega);
-        const double sin_omega = std::sin(omega);
-        pos(0) = x_orb * cos_omega - y_orb * cos_i * sin_omega;
-        pos(1) = x_orb * sin_omega + y_orb * cos_i * cos_omega;
-        pos(2) = y_orb * sin_i;
-    }
-
-    clock_bias = eph.af0 + eph.af1 * tc + eph.af2 * tc * tc;
-    clock_bias += kRelativityF * eph.e * eph.sqrt_a * sin_e;
-    // TGD NOT applied here — it's applied in SPP only (DD cancels TGD)
-    // clock_bias -= eph.tgd;
-    clock_drift = eph.af1 + 2.0 * eph.af2 * tc;
-    return true;
-}
-
-void computeGlonassDynamics(const double* state, double* derivative, const Vector3d& acceleration) {
-    const double r2 = state[0] * state[0] + state[1] * state[1] + state[2] * state[2];
-    if (r2 <= 0.0) {
-        for (int i = 0; i < 6; ++i) derivative[i] = 0.0;
-        return;
-    }
-
-    const double r3 = r2 * std::sqrt(r2);
-    const double omg2 = kGlonassEarthRotationRate * kGlonassEarthRotationRate;
-    const double a = 1.5 * kGlonassJ2 * kGlonassEarthMu *
-                     (kGlonassEarthRadius * kGlonassEarthRadius) / (r2 * r3);
-    const double b = 5.0 * state[2] * state[2] / r2;
-    const double c = -kGlonassEarthMu / r3 - a * (1.0 - b);
-
-    derivative[0] = state[3];
-    derivative[1] = state[4];
-    derivative[2] = state[5];
-    derivative[3] = (c + omg2) * state[0] + 2.0 * kGlonassEarthRotationRate * state[4] + acceleration(0);
-    derivative[4] = (c + omg2) * state[1] - 2.0 * kGlonassEarthRotationRate * state[3] + acceleration(1);
-    derivative[5] = (c - 2.0 * a) * state[2] + acceleration(2);
-}
-
-void propagateGlonassOrbit(double dt, double* state, const Vector3d& acceleration) {
-    double k1[6], k2[6], k3[6], k4[6], work[6];
-
-    computeGlonassDynamics(state, k1, acceleration);
-    for (int i = 0; i < 6; ++i) work[i] = state[i] + k1[i] * dt / 2.0;
-    computeGlonassDynamics(work, k2, acceleration);
-    for (int i = 0; i < 6; ++i) work[i] = state[i] + k2[i] * dt / 2.0;
-    computeGlonassDynamics(work, k3, acceleration);
-    for (int i = 0; i < 6; ++i) work[i] = state[i] + k3[i] * dt;
-    computeGlonassDynamics(work, k4, acceleration);
-
-    for (int i = 0; i < 6; ++i) {
-        state[i] += (k1[i] + 2.0 * k2[i] + 2.0 * k3[i] + k4[i]) * dt / 6.0;
-    }
-}
-
-bool computeGlonassState(const Ephemeris& eph,
-                         const GNSSTime& time,
-                         Vector3d& pos,
-                         Vector3d& vel,
-                         double& clock_bias,
-                         double& clock_drift) {
-    if (!eph.valid) {
-        return false;
-    }
-
-    double state[6] = {
-        eph.glonass_position(0), eph.glonass_position(1), eph.glonass_position(2),
-        eph.glonass_velocity(0), eph.glonass_velocity(1), eph.glonass_velocity(2),
-    };
-
-    double t = time - eph.toe;
-    clock_bias = -eph.glonass_taun + eph.glonass_gamn * t;
-    clock_drift = eph.glonass_gamn;
-
-    for (double step = t < 0.0 ? -kGlonassIntegrationStep : kGlonassIntegrationStep;
-         std::abs(t) > 1e-9;
-         t -= step) {
-        if (std::abs(t) < kGlonassIntegrationStep) {
-            step = t;
-        }
-        propagateGlonassOrbit(step, state, eph.glonass_acceleration);
-    }
-
-    pos = Vector3d(state[0], state[1], state[2]);
-    vel = Vector3d(state[3], state[4], state[5]);
-    return true;
-}
-
-bool computeSbasState(const Ephemeris& eph,
-                      const GNSSTime& time,
-                      Vector3d& pos,
-                      Vector3d& vel,
-                      double& clock_bias,
-                      double& clock_drift) {
-    if (!eph.valid) {
-        return false;
-    }
-
-    // SBAS GEO ephemerides are propagated as a polynomial in the ECEF frame
-    // (RTKLIB seph2pos): position and clock are evaluated at the time offset
-    // from the reference epoch, matching the broadcast element set.
-    const double t = time - eph.toe;
-    pos = eph.glonass_position +
-          eph.glonass_velocity * t +
-          eph.glonass_acceleration * (t * t / 2.0);
-    vel = eph.glonass_velocity + eph.glonass_acceleration * t;
-    clock_bias = eph.af0 + eph.af1 * t;
-    clock_drift = eph.af1;
-    return true;
-}
-
-void ecefToGeodetic(const Vector3d& ecef, double& lat, double& lon, double& h) {
-    lon = std::atan2(ecef(1), ecef(0));
-    const double p = std::sqrt(ecef(0) * ecef(0) + ecef(1) * ecef(1));
-    lat = std::atan2(ecef(2), p * (1.0 - kWgs84E2));
-
-    for (int i = 0; i < 8; ++i) {
-        const double sin_lat = std::sin(lat);
-        const double n = kWgs84A / std::sqrt(1.0 - kWgs84E2 * sin_lat * sin_lat);
-        h = p / std::cos(lat) - n;
-        lat = std::atan2(ecef(2), p * (1.0 - kWgs84E2 * n / (n + h)));
-    }
-
-    const double sin_lat = std::sin(lat);
-    const double n = kWgs84A / std::sqrt(1.0 - kWgs84E2 * sin_lat * sin_lat);
-    h = p / std::cos(lat) - n;
-}
-
-} // namespace
+using namespace navigation_internal;
 
 bool Ephemeris::calculateSatelliteState(const GNSSTime& time,
                                        Vector3d& pos,
@@ -1443,2595 +436,4 @@ double TroposphereModel::calculateDelay(const GeodeticCoord& user_pos,
     
     return Zd * mapping;
 }
-
-bool PreciseProducts::loadSP3File(const std::string& filename) {
-    std::ifstream input(filename);
-    if (!input.is_open()) {
-        return false;
-    }
-
-    GNSSTime current_time;
-    bool have_epoch = false;
-    bool loaded_any = false;
-    std::string line;
-    while (std::getline(input, line)) {
-        if (line.empty()) {
-            continue;
-        }
-        if (line[0] == '*') {
-            have_epoch = parseSp3EpochLine(line, current_time);
-            continue;
-        }
-        if (!have_epoch || line[0] != 'P') {
-            continue;
-        }
-
-        std::istringstream stream(line.substr(1));
-        std::string sat_token;
-        double x_km = 0.0;
-        double y_km = 0.0;
-        double z_km = 0.0;
-        double clock_microseconds = 0.0;
-        if (!(stream >> sat_token >> x_km >> y_km >> z_km >> clock_microseconds)) {
-            continue;
-        }
-
-        SatelliteId satellite;
-        if (!parseSatelliteToken(sat_token, satellite)) {
-            continue;
-        }
-
-        PreciseOrbitClock sample;
-        sample.satellite = satellite;
-        sample.time = current_time;
-        sample.position = Vector3d(x_km * 1000.0, y_km * 1000.0, z_km * 1000.0);
-        sample.position_valid = std::isfinite(sample.position.x()) &&
-            std::isfinite(sample.position.y()) &&
-            std::isfinite(sample.position.z());
-        if (std::isfinite(clock_microseconds) && std::abs(clock_microseconds) < 999999.0) {
-            sample.clock_bias = clock_microseconds * 1e-6;
-            sample.clock_valid = true;
-        }
-        addOrbitClock(sample);
-        loaded_any = true;
-    }
-    return loaded_any;
-}
-
-bool PreciseProducts::loadClockFile(const std::string& filename) {
-    std::ifstream input(filename);
-    if (!input.is_open()) {
-        return false;
-    }
-
-    bool in_header = true;
-    bool loaded_any = false;
-    std::string line;
-    while (std::getline(input, line)) {
-        if (in_header) {
-            if (line.find("END OF HEADER") != std::string::npos) {
-                in_header = false;
-            }
-            continue;
-        }
-        if (line.size() < 2 || line[0] != 'A' || line[1] != 'S') {
-            continue;
-        }
-
-        std::istringstream stream(line);
-        std::string record_type;
-        std::string sat_token;
-        if (!(stream >> record_type >> sat_token)) {
-            continue;
-        }
-
-        SatelliteId satellite;
-        if (!parseSatelliteToken(sat_token, satellite)) {
-            continue;
-        }
-
-        GNSSTime current_time;
-        if (!parseClockEpochTokens(stream, current_time)) {
-            continue;
-        }
-
-        int value_count = 0;
-        double clock_bias = 0.0;
-        double clock_sigma = 0.0;
-        if (!(stream >> value_count >> clock_bias)) {
-            continue;
-        }
-        if (value_count > 1) {
-            stream >> clock_sigma;
-        }
-
-        PreciseOrbitClock sample;
-        sample.satellite = satellite;
-        sample.time = current_time;
-        sample.clock_bias = clock_bias;
-        sample.clock_sigma = clock_sigma;
-        sample.clock_valid = std::isfinite(clock_bias);
-        addOrbitClock(sample);
-        loaded_any = true;
-    }
-    return loaded_any;
-}
-
-void PreciseProducts::addOrbitClock(const PreciseOrbitClock& data) {
-    auto& entries = orbit_clock_data[data.satellite];
-    auto it = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        data.time,
-        [](const PreciseOrbitClock& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-
-    if (it != entries.end() && it->time == data.time) {
-        if (data.position_valid) {
-            it->position = data.position;
-            it->velocity = data.velocity;
-            it->position_sigma = data.position_sigma;
-            it->position_valid = true;
-        }
-        if (data.clock_valid) {
-            it->clock_bias = data.clock_bias;
-            it->clock_drift = data.clock_drift;
-            it->clock_sigma = data.clock_sigma;
-            it->clock_valid = true;
-        }
-        return;
-    }
-
-    entries.insert(it, data);
-}
-
-namespace {
-
-// RTKLIB-style Neville's algorithm for polynomial interpolation. Returns the
-// interpolated value at t=0 given samples at offsets x[i] with values y[i].
-// y[] is modified in place. Numerically stable for closely-spaced samples.
-double nevillePolynomial(const std::vector<double>& x,
-                         std::vector<double>& y,
-                         int n) {
-    for (int j = 1; j < n; ++j) {
-        for (int i = 0; i < n - j; ++i) {
-            const double denom = x[i + j] - x[i];
-            if (std::abs(denom) < 1e-12) {
-                return y[i];
-            }
-            y[i] = (x[i + j] * y[i] - x[i] * y[i + 1]) / denom;
-        }
-    }
-    return y[0];
-}
-
-constexpr int kLagrangeOrder = 10;  ///< number of SP3 samples (degree 9)
-
-}  // namespace
-
-bool PreciseProducts::interpolateOrbitClock(const SatelliteId& sat,
-                                            const GNSSTime& time,
-                                            Vector3d& position,
-                                            Vector3d& velocity,
-                                            double& clock_bias,
-                                            double& clock_drift) const {
-    const auto sat_it = orbit_clock_data.find(sat);
-    if (sat_it == orbit_clock_data.end() || sat_it->second.empty()) {
-        return false;
-    }
-
-    const auto& entries = sat_it->second;
-
-    // Build position- and clock-sample index vectors so polynomial
-    // interpolation can pick a contiguous window of N samples around the
-    // query time. SP3 (orbit) and CLK rows are stored in the same sorted
-    // list; position_valid and clock_valid flag which records contribute
-    // to each grid.
-    std::vector<int> pos_indices;
-    std::vector<int> clk_indices;
-    int pos_at_or_before = -1;  // last position-valid index with time <= query
-    int clk_at_or_before = -1;
-    pos_indices.reserve(entries.size());
-    clk_indices.reserve(entries.size());
-    for (int i = 0; i < static_cast<int>(entries.size()); ++i) {
-        if (entries[i].position_valid) {
-            pos_indices.push_back(i);
-            if (entries[i].time <= time) {
-                pos_at_or_before = static_cast<int>(pos_indices.size()) - 1;
-            }
-        }
-        if (entries[i].clock_valid) {
-            clk_indices.push_back(i);
-            if (entries[i].time <= time) {
-                clk_at_or_before = static_cast<int>(clk_indices.size()) - 1;
-            }
-        }
-    }
-
-    if (pos_indices.empty()) {
-        return false;
-    }
-
-    // Pick an N-sample window centred on the query, clamped to the array
-    // ends so the same code path serves boundary queries with extrapolation
-    // (RTKLIB pephpos does the same).
-    auto interpolateLagrange = [&entries, &time](
-                                    const std::vector<int>& indices,
-                                    int at_or_before,
-                                    auto&& accessor,
-                                    auto& out_value,
-                                    auto& out_rate) -> bool {
-        const int n_avail = static_cast<int>(indices.size());
-        if (n_avail == 0) {
-            return false;
-        }
-        const int n_use = std::min(kLagrangeOrder, n_avail);
-        int start = at_or_before - (n_use - 1) / 2;
-        if (start < 0) start = 0;
-        if (start + n_use > n_avail) start = n_avail - n_use;
-        if (start < 0) start = 0;
-
-        // Reference the times relative to the first picked sample to keep
-        // the polynomial coefficients well-scaled (15-min spacing × 9 ≈ 8100 s
-        // dynamic range; relative differences fit cleanly in double).
-        const GNSSTime& t_ref = entries[indices[start]].time;
-        const double dt_query = time - t_ref;
-        if (std::abs(dt_query) > kPreciseInterpolationGapSeconds * n_use) {
-            return false;  // sanity guard: query far outside sample span
-        }
-
-        std::vector<double> xs;
-        xs.reserve(n_use);
-        for (int k = 0; k < n_use; ++k) {
-            xs.push_back(entries[indices[start + k]].time - t_ref - dt_query);
-        }
-
-        // For derivative computation: evaluate the same polynomial at query+h
-        // (i.e. shift the x-axis by -h so the new origin is at original x=h).
-        // dp/dt = (P(h) - P(0))/h.
-        const double h_deriv = 1.0;
-        std::vector<double> xs_deriv = xs;
-        for (auto& x : xs_deriv) x -= h_deriv;
-
-        using value_t = decltype(accessor(entries[indices[0]]));
-        if constexpr (std::is_same_v<value_t, Vector3d>) {
-            for (int axis = 0; axis < 3; ++axis) {
-                std::vector<double> ys_value;
-                std::vector<double> ys_deriv;
-                ys_value.reserve(n_use);
-                ys_deriv.reserve(n_use);
-                for (int k = 0; k < n_use; ++k) {
-                    const double v = accessor(entries[indices[start + k]])(axis);
-                    ys_value.push_back(v);
-                    ys_deriv.push_back(v);
-                }
-                out_value(axis) = nevillePolynomial(xs, ys_value, n_use);
-                const double v_shift = nevillePolynomial(xs_deriv, ys_deriv, n_use);
-                out_rate(axis) = (v_shift - out_value(axis)) / h_deriv;
-            }
-        } else {
-            std::vector<double> ys_value;
-            std::vector<double> ys_deriv;
-            ys_value.reserve(n_use);
-            ys_deriv.reserve(n_use);
-            for (int k = 0; k < n_use; ++k) {
-                const double v = accessor(entries[indices[start + k]]);
-                ys_value.push_back(v);
-                ys_deriv.push_back(v);
-            }
-            out_value = nevillePolynomial(xs, ys_value, n_use);
-            const double v_shift = nevillePolynomial(xs_deriv, ys_deriv, n_use);
-            out_rate = (v_shift - out_value) / h_deriv;
-        }
-        return true;
-    };
-
-    if (!interpolateLagrange(
-            pos_indices, pos_at_or_before,
-            [](const PreciseOrbitClock& e) -> Vector3d { return e.position; },
-            position, velocity)) {
-        return false;
-    }
-
-    if (!interpolateLagrange(
-            clk_indices, clk_at_or_before,
-            [](const PreciseOrbitClock& e) -> double { return e.clock_bias; },
-            clock_bias, clock_drift)) {
-        clock_bias = 0.0;
-        clock_drift = 0.0;
-    }
-    return true;
-}
-
-bool PreciseProducts::hasData(const SatelliteId& sat, const GNSSTime& time) const {
-    Vector3d position;
-    Vector3d velocity;
-    double clock_bias = 0.0;
-    double clock_drift = 0.0;
-    return interpolateOrbitClock(sat, time, position, velocity, clock_bias, clock_drift);
-}
-
-void PreciseProducts::clear() {
-    orbit_clock_data.clear();
-}
-
-void SSRProducts::addCorrection(const SSROrbitClockCorrection& correction) {
-    auto& entries = orbit_clock_corrections[correction.satellite];
-    auto lower = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        correction.time,
-        [](const SSROrbitClockCorrection& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-    auto upper = std::upper_bound(
-        lower,
-        entries.end(),
-        correction.time,
-        [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-            return lhs < rhs.time;
-        });
-
-    for (auto it = lower; it != upper; ++it) {
-        if (!sameCorrectionVariant(*it, correction)) {
-            continue;
-        }
-        mergeCorrectionFields(*it, correction);
-        return;
-    }
-
-    SSROrbitClockCorrection to_insert = correction;
-    if (to_insert.clock_valid && to_insert.clock_network_id == 0) {
-        to_insert.mrtklib_base_clock_correction_m =
-            to_insert.clock_correction_m;
-        to_insert.mrtklib_base_clock_reference_time =
-            to_insert.clock_reference_time.week != 0
-                ? to_insert.clock_reference_time
-                : to_insert.time;
-        to_insert.mrtklib_base_clock_valid = true;
-    }
-    promoteClaslibBaseClockFields(to_insert);
-    entries.insert(upper, to_insert);
-}
-
-void SSRProducts::addCorrections(const std::vector<SSROrbitClockCorrection>& corrections) {
-    if (corrections.empty()) {
-        return;
-    }
-
-    struct CorrectionKey {
-        GNSSTime time;
-        int atmos_network_id = 0;
-        int bias_network_id = 0;
-
-        bool operator<(const CorrectionKey& other) const {
-            if (time < other.time) {
-                return true;
-            }
-            if (other.time < time) {
-                return false;
-            }
-            if (atmos_network_id != other.atmos_network_id) {
-                return atmos_network_id < other.atmos_network_id;
-            }
-            return bias_network_id < other.bias_network_id;
-        }
-    };
-
-    struct PendingSatelliteCorrections {
-        std::vector<SSROrbitClockCorrection> entries;
-        std::map<CorrectionKey, size_t> index_by_key;
-    };
-
-    std::map<SatelliteId, PendingSatelliteCorrections> pending;
-    for (const auto& [satellite, entries] : orbit_clock_corrections) {
-        PendingSatelliteCorrections& sat_pending = pending[satellite];
-        sat_pending.entries = entries;
-        for (size_t index = 0; index < sat_pending.entries.size(); ++index) {
-            const SSROrbitClockCorrection& entry = sat_pending.entries[index];
-            const CorrectionKey key{entry.time, entry.atmos_network_id, entry.bias_network_id};
-            sat_pending.index_by_key.emplace(key, index);
-        }
-    }
-
-    for (const SSROrbitClockCorrection& correction : corrections) {
-        PendingSatelliteCorrections& sat_pending = pending[correction.satellite];
-        const CorrectionKey key{
-            correction.time,
-            correction.atmos_network_id,
-            correction.bias_network_id,
-        };
-        const auto index_it = sat_pending.index_by_key.find(key);
-        if (index_it != sat_pending.index_by_key.end()) {
-            mergeCorrectionFields(sat_pending.entries[index_it->second], correction);
-            continue;
-        }
-        sat_pending.index_by_key.emplace(key, sat_pending.entries.size());
-        sat_pending.entries.push_back(correction);
-    }
-
-    orbit_clock_corrections.clear();
-    for (auto& [satellite, sat_pending] : pending) {
-        std::stable_sort(
-            sat_pending.entries.begin(),
-            sat_pending.entries.end(),
-            [](const SSROrbitClockCorrection& lhs,
-               const SSROrbitClockCorrection& rhs) {
-                return lhs.time < rhs.time;
-            });
-        orbit_clock_corrections.emplace(satellite, std::move(sat_pending.entries));
-    }
-}
-
-bool SSRProducts::interpolateCorrection(const SatelliteId& sat,
-                                        const GNSSTime& time,
-                                        Vector3d& orbit_correction_ecef,
-                                        double& clock_correction_m,
-                                        double* ura_sigma_m,
-                                        std::map<uint8_t, double>* code_bias_m,
-                                        std::map<uint8_t, double>* phase_bias_m,
-                                        std::map<std::string, std::string>* atmos_tokens,
-                                        GNSSTime* atmos_reference_time,
-                                        GNSSTime* phase_bias_reference_time,
-                                        GNSSTime* clock_reference_time,
-                                        int preferred_network_id,
-                                        int* orbit_iode,
-                                        std::map<uint8_t, int>* phase_bias_discnt,
-                                        SSRCorrectionStatus* status,
-                                        bool allow_future_samples,
-                                        double* base_clock_correction_m,
-                                        bool* base_clock_valid,
-                                        SSRClockSelectionPolicy clock_selection_policy,
-                                        std::map<uint8_t, double>* code_bias_rtklib_m,
-                                        std::map<uint8_t, double>* phase_bias_rtklib_m) const {
-    if (status != nullptr) {
-        *status = SSRCorrectionStatus{};
-    }
-    const auto sat_it = orbit_clock_corrections.find(sat);
-    if (sat_it == orbit_clock_corrections.end() || sat_it->second.empty()) {
-        return false;
-    }
-    if (orbit_iode != nullptr) {
-        *orbit_iode = -1;
-    }
-    orbit_correction_ecef.setZero();
-    clock_correction_m = 0.0;
-    if (ura_sigma_m != nullptr) {
-        *ura_sigma_m = 0.0;
-    }
-    if (code_bias_m != nullptr) {
-        code_bias_m->clear();
-    }
-    if (phase_bias_m != nullptr) {
-        phase_bias_m->clear();
-    }
-    if (code_bias_rtklib_m != nullptr) {
-        code_bias_rtklib_m->clear();
-    }
-    if (phase_bias_rtklib_m != nullptr) {
-        phase_bias_rtklib_m->clear();
-    }
-    if (phase_bias_discnt != nullptr) {
-        phase_bias_discnt->clear();
-    }
-    if (atmos_tokens != nullptr) {
-        atmos_tokens->clear();
-    }
-    if (atmos_reference_time != nullptr) {
-        *atmos_reference_time = GNSSTime();
-    }
-    if (phase_bias_reference_time != nullptr) {
-        *phase_bias_reference_time = GNSSTime();
-    }
-    if (clock_reference_time != nullptr) {
-        *clock_reference_time = GNSSTime();
-    }
-
-    const auto& entries = sat_it->second;
-    const bool galileo_clock_withdrawal_active =
-        sat.system == GNSSSystem::Galileo &&
-        (clock_selection_policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold ||
-         pppEnvOverrides().clas_code_row_full_prc);
-    if (galileo_clock_withdrawal_active) {
-        const SSROrbitClockCorrection* newest_clock_event = nullptr;
-        auto event_end = std::upper_bound(
-            entries.begin(),
-            entries.end(),
-            time,
-            [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                return lhs < rhs.time;
-            });
-        while (event_end != entries.begin()) {
-            const SSROrbitClockCorrection& entry = *--event_end;
-            if (time - entry.time > kClaslibMaxClockAgeFromObsSeconds) {
-                break;
-            }
-            if (!entry.clock_withdrawn && !entry.mrtklib_base_clock_valid) {
-                continue;
-            }
-            if (newest_clock_event == nullptr ||
-                entry.time > newest_clock_event->time ||
-                (entry.time == newest_clock_event->time &&
-                 entry.clock_withdrawn && !newest_clock_event->clock_withdrawn)) {
-                newest_clock_event = &entry;
-            }
-        }
-        if (newest_clock_event != nullptr && newest_clock_event->clock_withdrawn) {
-            if (base_clock_correction_m != nullptr) {
-                *base_clock_correction_m = 0.0;
-            }
-            if (base_clock_valid != nullptr) {
-                *base_clock_valid = false;
-            }
-            if (status != nullptr) {
-                status->clock_withdrawn = true;
-                status->clock_reference_time = newest_clock_event->time;
-            }
-            return false;
-        }
-    }
-    auto lower = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        time,
-        [](const SSROrbitClockCorrection& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-    auto upper = std::upper_bound(
-        lower,
-        entries.end(),
-        time,
-        [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-            return lhs < rhs.time;
-        });
-
-    const auto biasScore = [&](const SSROrbitClockCorrection& entry, bool phase) -> int {
-        const bool valid =
-            phase
-                ? (entry.phase_bias_valid && !entry.phase_bias_m.empty())
-                : (entry.code_bias_valid && !entry.code_bias_m.empty());
-        if (!valid) {
-            return -1;
-        }
-        const size_t signal_count = phase ? entry.phase_bias_m.size() : entry.code_bias_m.size();
-        if (preferred_network_id > 0) {
-            if (entry.bias_network_id == preferred_network_id) {
-                return 300000 + static_cast<int>(signal_count);
-            }
-            if (entry.bias_network_id == 0) {
-                return 200000 + static_cast<int>(signal_count);
-            }
-            return 100000 + static_cast<int>(signal_count);
-        }
-        if (entry.bias_network_id == 0) {
-            return 200000 + static_cast<int>(signal_count);
-        }
-        return 100000 + static_cast<int>(signal_count);
-    };
-
-    const bool qzss_atmos_fix =
-        pppEnvOverrides().clas_qzss_s_prn_fix && sat.system == GNSSSystem::QZSS;
-
-    // Compact SSR orbit subtype rows form a constellation-wide bank.  For the
-    // legacy S120 -> J01 identity, CLASLIB clears the SSR orbit when a newer
-    // bank omits S120 instead of carrying the preceding value. J02/J03 also
-    // have direct compact identities, so applying the S-PRN withdrawal rule to
-    // them would conflate direct J02/J03 rows with legacy S121/S122 aliases.
-    // Keep this behind the explicit QZSS parity gate so ordinary SSR/MADOCA
-    // selection retains the historical per-satellite hold.
-    const auto qzssOrbitWithdrawnByNewBank =
-        [&](const SSROrbitClockCorrection* source) -> bool {
-        if (!qzss_atmos_fix || sat.prn != 1 || source == nullptr ||
-            !source->orbit_valid) {
-            return false;
-        }
-        const GNSSTime source_time = orbitReferenceTime(*source);
-        GNSSTime newest_bank_time;
-        bool have_newest_bank = false;
-        for (const auto& [candidate_sat, candidate_entries] : orbit_clock_corrections) {
-            if (candidate_sat.system != GNSSSystem::QZSS) {
-                continue;
-            }
-            auto candidate_end = std::upper_bound(
-                candidate_entries.begin(),
-                candidate_entries.end(),
-                time,
-                [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                    return lhs < rhs.time;
-                });
-            while (candidate_end != candidate_entries.begin()) {
-                const SSROrbitClockCorrection& candidate = *--candidate_end;
-                const GNSSTime bank_time = orbitReferenceTime(candidate);
-                if (bank_time > time) {
-                    continue;
-                }
-                if (have_newest_bank && candidate.time < newest_bank_time) {
-                    break;
-                }
-                if (!candidate.orbit_valid) {
-                    continue;
-                }
-                if (!have_newest_bank || bank_time > newest_bank_time) {
-                    newest_bank_time = bank_time;
-                    have_newest_bank = true;
-                }
-            }
-        }
-        return have_newest_bank && source_time < newest_bank_time;
-    };
-
-    const auto atmosScore = [&](const SSROrbitClockCorrection& entry) -> int {
-        if (!entry.atmos_valid || entry.atmos_tokens.empty()) {
-            return -1;
-        }
-        int score = static_cast<int>(entry.atmos_tokens.size());
-        if (entry.atmos_network_id > 0) {
-            score += 100000;
-        }
-        if (preferred_network_id > 0) {
-            if (entry.atmos_network_id == preferred_network_id) {
-                score += 200000;
-            } else if (entry.atmos_network_id == 0) {
-                score += 50000;
-            }
-            // QZSS regional atmos (e.g. network 19 on S120-S122 rows) is valid near
-            // Japan grids. Network-1 polynomial coeffs for J01-J03 are only meaningful
-            // near network-1 grids (Okinawa); applying them at Chiba pierce points
-            // yields ~30 TECU STEC error.
-            if (qzss_atmos_fix && preferred_network_id == 1 &&
-                entry.atmos_network_id > 1) {
-                score += 300000;
-            }
-        }
-        return score;
-    };
-
-    const auto pickBestAtmosAmong = [&](size_t begin, size_t end)
-        -> const SSROrbitClockCorrection* {
-        bool have_regional_qzss_atmos = false;
-        if (qzss_atmos_fix) {
-            for (size_t index = begin; index < end; ++index) {
-                if (entries[index].atmos_valid &&
-                    entries[index].atmos_network_id > 1) {
-                    have_regional_qzss_atmos = true;
-                    break;
-                }
-            }
-        }
-        const SSROrbitClockCorrection* best = nullptr;
-        int best_score = -1;
-        for (size_t index = begin; index < end; ++index) {
-            if (!entries[index].atmos_valid || entries[index].atmos_tokens.empty()) {
-                continue;
-            }
-            if (have_regional_qzss_atmos &&
-                entries[index].atmos_network_id <= 1) {
-                continue;
-            }
-            const int score = atmosScore(entries[index]);
-            if (score > best_score) {
-                best_score = score;
-                best = &entries[index];
-            }
-        }
-        return best;
-    };
-
-    const auto findQzssHeldCompactAtmos = [&](size_t scan_end_index)
-        -> const SSROrbitClockCorrection* {
-        const SSROrbitClockCorrection* held_preferred = nullptr;
-        const SSROrbitClockCorrection* held_net19 = nullptr;
-        GNSSTime held_preferred_time;
-        GNSSTime held_net19_time;
-        size_t group_end = scan_end_index;
-        while (group_end > 0) {
-            const size_t group_last = group_end - 1;
-            const GNSSTime group_time = entries[group_last].time;
-            if (time - group_time > kClasStecValidAgeSeconds + 1e-9) {
-                break;
-            }
-            size_t group_begin = group_last;
-            while (group_begin > 0 && entries[group_begin - 1].time == group_time) {
-                --group_begin;
-            }
-            for (size_t index = group_begin; index < group_end; ++index) {
-                const SSROrbitClockCorrection& entry = entries[index];
-                if (!entry.atmos_valid || entry.atmos_tokens.empty()) {
-                    continue;
-                }
-                if (preferred_network_id > 0 &&
-                    entry.atmos_network_id == preferred_network_id &&
-                    (held_preferred == nullptr || entry.time > held_preferred_time)) {
-                    held_preferred = &entry;
-                    held_preferred_time = entry.time;
-                }
-                if (entry.atmos_network_id == 19 &&
-                    (held_net19 == nullptr || entry.time > held_net19_time)) {
-                    held_net19 = &entry;
-                    held_net19_time = entry.time;
-                }
-            }
-            group_end = group_begin;
-        }
-        // Regional compact network 19 must win over epoch-preferred network 1 when both
-        // are held within age (ST9-only epochs still carry net-1 decoy polynomials).
-        if (held_net19 != nullptr) {
-            return held_net19;
-        }
-        if (held_preferred != nullptr) {
-            return held_preferred;
-        }
-        if (scan_end_index > 0) {
-            const size_t group_last = scan_end_index - 1;
-            const GNSSTime group_time = entries[group_last].time;
-            if (time - group_time <= kClasStecValidAgeSeconds + 1e-9) {
-                size_t group_begin = group_last;
-                while (group_begin > 0 && entries[group_begin - 1].time == group_time) {
-                    --group_begin;
-                }
-                return pickBestAtmosAmong(group_begin, scan_end_index);
-            }
-        }
-        return nullptr;
-    };
-
-    const auto isQzssDedicatedBiasBank = [&](const SSROrbitClockCorrection& entry, bool phase) -> bool {
-        return isQzssDedicatedBiasBankEntry(entry, phase);
-    };
-
-    const auto findQzssHeldDedicatedBias = [&](bool phase, size_t scan_end_index)
-        -> const SSROrbitClockCorrection* {
-        const SSROrbitClockCorrection* held = nullptr;
-        GNSSTime held_time;
-        size_t group_end = scan_end_index;
-        while (group_end > 0) {
-            const size_t group_last = group_end - 1;
-            const GNSSTime group_time = entries[group_last].time;
-            if (time - group_time > kClasStecValidAgeSeconds + 1e-9) {
-                break;
-            }
-            size_t group_begin = group_last;
-            while (group_begin > 0 && entries[group_begin - 1].time == group_time) {
-                --group_begin;
-            }
-            for (size_t index = group_begin; index < group_end; ++index) {
-                const SSROrbitClockCorrection& entry = entries[index];
-                if (!isQzssDedicatedBiasBank(entry, phase)) {
-                    continue;
-                }
-                // CLASLIB exposes a newly decoded compact bias bank only
-                // after its 15 s reception delay.  The QZSS-specific path
-                // bypasses scanBackwardBias(), so enforce the same causal
-                // boundary here instead of consuming the bank at its tagged
-                // epoch.
-                if (time - entry.time < kMrtklibBiasReceptionLagSeconds) {
-                    continue;
-                }
-                if (held == nullptr || entry.time > held_time) {
-                    held = &entry;
-                    held_time = entry.time;
-                }
-            }
-            group_end = group_begin;
-        }
-        return held;
-    };
-
-    const auto findQzssHeldPhaseBiasForNetwork = [&](int service_network_id, size_t scan_end_index)
-        -> const SSROrbitClockCorrection* {
-        return findQzssHeldPhaseBiasForServiceNetwork(
-            entries, time, service_network_id, scan_end_index);
-    };
-
-    const auto pickAtmosForTime = [&](const GNSSTime& atmos_time)
-        -> const SSROrbitClockCorrection* {
-        const double max_age =
-            qzss_atmos_fix ? kClasStecValidAgeSeconds : kPreciseInterpolationGapSeconds;
-        if (time - atmos_time > max_age + 1e-9 || atmos_time - time > 1e-9) {
-            return nullptr;
-        }
-        const auto group_lower = std::lower_bound(
-            entries.begin(),
-            entries.end(),
-            atmos_time,
-            [](const SSROrbitClockCorrection& lhs, const GNSSTime& rhs) {
-                return lhs.time < rhs;
-            });
-        if (group_lower == entries.end() || group_lower->time != atmos_time) {
-            return nullptr;
-        }
-        const auto group_upper = std::upper_bound(
-            group_lower,
-            entries.end(),
-            atmos_time,
-            [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                return lhs < rhs.time;
-            });
-        const size_t group_begin = static_cast<size_t>(group_lower - entries.begin());
-        const size_t group_end = static_cast<size_t>(group_upper - entries.begin());
-        return pickBestAtmosAmong(group_begin, group_end);
-    };
-
-    const auto pickQzssHeldAtmos = [&]() -> const SSROrbitClockCorrection* {
-        size_t scan_end = entries.size();
-        if (lower != entries.end()) {
-            scan_end = static_cast<size_t>(lower - entries.begin()) + 1;
-        }
-        return findQzssHeldCompactAtmos(scan_end);
-    };
-
-    const auto scanBackwardBias = [&](bool phase) -> const SSROrbitClockCorrection* {
-        const auto scan = [&](int required_network_id)
-            -> const SSROrbitClockCorrection* {
-            size_t group_end = static_cast<size_t>(lower - entries.begin());
-            while (group_end > 0) {
-                const size_t group_last = group_end - 1;
-                const GNSSTime group_time = entries[group_last].time;
-                if (std::abs(group_time - time) >
-                    kPreciseInterpolationGapSeconds) {
-                    break;
-                }
-                size_t group_begin = group_last;
-                while (group_begin > 0 &&
-                       entries[group_begin - 1].time == group_time) {
-                    --group_begin;
-                }
-                if (clock_selection_policy ==
-                        SSRClockSelectionPolicy::MrtklibLiteralBaseHold &&
-                    time - group_time < kMrtklibBiasReceptionLagSeconds) {
-                    group_end = group_begin;
-                    continue;
-                }
-
-                const SSROrbitClockCorrection* best = nullptr;
-                int best_score = -1;
-                for (size_t index = group_begin; index < group_end; ++index) {
-                    if (required_network_id >= 0 &&
-                        entries[index].bias_network_id != required_network_id) {
-                        continue;
-                    }
-                    const int score = biasScore(entries[index], phase);
-                    if (score > best_score) {
-                        best_score = score;
-                        best = &entries[index];
-                    }
-                }
-                if (best != nullptr) {
-                    return best;
-                }
-                group_end = group_begin;
-            }
-            return nullptr;
-        };
-        if (clock_selection_policy ==
-            SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-            if (preferred_network_id > 0) {
-                if (const SSROrbitClockCorrection* selected =
-                        scan(preferred_network_id)) {
-                    return selected;
-                }
-            }
-            if (const SSROrbitClockCorrection* common = scan(0)) {
-                return common;
-            }
-        }
-        return scan(-1);
-    };
-
-    if (lower != entries.end() && lower->time == time) {
-        const size_t lower_index = static_cast<size_t>(lower - entries.begin());
-        size_t exact_begin = lower_index;
-        while (exact_begin > 0 && entries[exact_begin - 1].time == time) {
-            --exact_begin;
-        }
-        size_t exact_end = lower_index;
-        while (exact_end < entries.size() && entries[exact_end].time == time) {
-            ++exact_end;
-        }
-
-        const auto findRecent = [&](const auto& picker, size_t scan_end_index)
-            -> const SSROrbitClockCorrection* {
-            size_t group_end = scan_end_index;
-            while (group_end > 0) {
-                const size_t group_last = group_end - 1;
-                const GNSSTime group_time = entries[group_last].time;
-                if (std::abs(group_time - time) > kPreciseInterpolationGapSeconds) {
-                    break;
-                }
-                size_t group_begin = group_last;
-                while (group_begin > 0 && entries[group_begin - 1].time == group_time) {
-                    --group_begin;
-                }
-                if (const SSROrbitClockCorrection* picked = picker(group_begin, group_end)) {
-                    return picked;
-                }
-                group_end = group_begin;
-            }
-            return nullptr;
-        };
-
-        const auto pickOrbit = [&](size_t begin, size_t end) -> const SSROrbitClockCorrection* {
-            for (size_t index = begin; index < end; ++index) {
-                if (entries[index].orbit_valid) {
-                    return &entries[index];
-                }
-            }
-            return nullptr;
-        };
-        const auto pickClock = [&](size_t begin, size_t end) -> const SSROrbitClockCorrection* {
-            const SSROrbitClockCorrection* fallback = nullptr;
-            for (size_t index = begin; index < end; ++index) {
-                if (!entries[index].clock_valid) {
-                    continue;
-                }
-                if (clasQzssBaseClockHygieneActive(sat) &&
-                    entries[index].clock_network_id > 0 &&
-                    !atmosEmbeddedClockRow(entries[index])) {
-                    return &entries[index];
-                }
-                if (fallback == nullptr) {
-                    fallback = &entries[index];
-                }
-            }
-            return fallback;
-        };
-        const auto pickUra = [&](size_t begin, size_t end) -> const SSROrbitClockCorrection* {
-            for (size_t index = begin; index < end; ++index) {
-                if (entries[index].ura_valid) {
-                    return &entries[index];
-                }
-            }
-            return nullptr;
-        };
-        const auto pickAtmos = [&](size_t begin, size_t end) -> const SSROrbitClockCorrection* {
-            if (qzss_atmos_fix) {
-                return pickBestAtmosAmong(begin, end);
-            }
-            const SSROrbitClockCorrection* best = nullptr;
-            int best_score = -1;
-            for (size_t index = begin; index < end; ++index) {
-                const int score = atmosScore(entries[index]);
-                if (score > best_score) {
-                    best_score = score;
-                    best = &entries[index];
-                }
-            }
-            return best;
-        };
-
-        const SSROrbitClockCorrection* atmos_source =
-            qzss_atmos_fix ? findQzssHeldCompactAtmos(exact_end)
-                           : findRecent(pickAtmos, exact_end);
-        int selected_network_id = preferred_network_id;
-        if (selected_network_id <= 0 && atmos_source != nullptr && atmos_source->atmos_network_id > 0) {
-            selected_network_id = atmos_source->atmos_network_id;
-        }
-        const auto pickBias = [&](bool phase) {
-            return [&, phase](size_t begin, size_t end) -> const SSROrbitClockCorrection* {
-                const SSROrbitClockCorrection* best = nullptr;
-                int best_score = -1;
-                for (size_t index = begin; index < end; ++index) {
-                    const SSROrbitClockCorrection& entry = entries[index];
-                    const bool valid =
-                        phase
-                            ? (entry.phase_bias_valid && !entry.phase_bias_m.empty())
-                            : (entry.code_bias_valid && !entry.code_bias_m.empty());
-                    if (!valid) {
-                        continue;
-                    }
-                    if (clock_selection_policy ==
-                            SSRClockSelectionPolicy::MrtklibLiteralBaseHold &&
-                        time - entry.time < kMrtklibBiasReceptionLagSeconds) {
-                        continue;
-                    }
-                    const size_t signal_count =
-                        phase ? entry.phase_bias_m.size() : entry.code_bias_m.size();
-                    int score = static_cast<int>(signal_count);
-                    if (selected_network_id > 0) {
-                        if (entry.bias_network_id == selected_network_id) {
-                            score += 300000;
-                        } else if (entry.bias_network_id == 0) {
-                            score += 200000;
-                        } else {
-                            score += 100000;
-                        }
-                    } else if (entry.bias_network_id == 0) {
-                        score += 200000;
-                    } else {
-                        score += 100000;
-                    }
-                    if (score > best_score) {
-                        best_score = score;
-                        best = &entry;
-                    }
-                }
-                return best;
-            };
-        };
-        const auto pickNetworkBias = [&](bool phase, int network_id) {
-            return [&, phase, network_id](size_t begin, size_t end)
-                -> const SSROrbitClockCorrection* {
-                const SSROrbitClockCorrection* best = nullptr;
-                size_t best_signal_count = 0;
-                for (size_t index = begin; index < end; ++index) {
-                    const SSROrbitClockCorrection& entry = entries[index];
-                    const bool valid = phase
-                        ? (entry.phase_bias_valid &&
-                           !entry.phase_bias_m.empty())
-                        : (entry.code_bias_valid &&
-                           !entry.code_bias_m.empty());
-                    if (!valid || entry.bias_network_id != network_id) {
-                        continue;
-                    }
-                    if (clock_selection_policy ==
-                            SSRClockSelectionPolicy::MrtklibLiteralBaseHold &&
-                        time - entry.time < kMrtklibBiasReceptionLagSeconds) {
-                        continue;
-                    }
-                    const size_t signal_count = phase
-                        ? entry.phase_bias_m.size()
-                        : entry.code_bias_m.size();
-                    if (best == nullptr || signal_count > best_signal_count) {
-                        best = &entry;
-                        best_signal_count = signal_count;
-                    }
-                }
-                return best;
-            };
-        };
-        const auto findRecentBias = [&](bool phase)
-            -> const SSROrbitClockCorrection* {
-            // A newer correction for another CLAS service area is not a
-            // refresh of the receiver's selected network. Search the entire
-            // hold window for that network first, then use the historical
-            // generic fallback only when no matching bank exists.
-            if (clock_selection_policy ==
-                SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-                if (selected_network_id > 0) {
-                    if (const SSROrbitClockCorrection* selected =
-                            findRecent(
-                                pickNetworkBias(phase, selected_network_id),
-                                exact_end)) {
-                        return selected;
-                    }
-                }
-                if (const SSROrbitClockCorrection* common =
-                        findRecent(pickNetworkBias(phase, 0), exact_end)) {
-                    return common;
-                }
-            }
-            return findRecent(pickBias(phase), exact_end);
-        };
-
-        const SSROrbitClockCorrection* orbit_source = findRecent(pickOrbit, exact_end);
-        if (qzssOrbitWithdrawnByNewBank(orbit_source)) {
-            orbit_source = nullptr;
-            if (status != nullptr) {
-                status->orbit_withdrawn = true;
-            }
-        }
-        const SSROrbitClockCorrection* clock_source = findRecent(pickClock, exact_end);
-        const SSROrbitClockCorrection* ura_source = findRecent(pickUra, exact_end);
-        const SSROrbitClockCorrection* code_source = nullptr;
-        if (code_bias_m != nullptr) {
-            code_source = qzss_atmos_fix
-                ? findQzssHeldDedicatedBias(false, exact_end)
-                : findRecentBias(false);
-        }
-        const SSROrbitClockCorrection* phase_source = nullptr;
-        if (phase_bias_m != nullptr && !qzss_atmos_fix) {
-            phase_source = findRecentBias(true);
-        }
-
-        if (orbit_source != nullptr && orbit_source->orbit_valid) {
-            orbit_correction_ecef = orbit_source->orbit_correction_ecef;
-            if (orbit_iode != nullptr) {
-                *orbit_iode = orbit_source->iode;
-            }
-            if (status != nullptr) {
-                status->orbit_valid = true;
-                status->orbit_reference_time = orbitReferenceTime(*orbit_source);
-                status->orbit_iode = orbit_source->iode;
-                status->ssr_orbit_iod = orbit_source->ssr_orbit_iod;
-            }
-        }
-        if (clock_source != nullptr && clock_source->clock_valid) {
-            clock_correction_m = clock_source->clock_correction_m;
-            assignBaseClockOutputs(*clock_source, clock_correction_m,
-                                   base_clock_correction_m, base_clock_valid);
-            if (clock_reference_time != nullptr) {
-                *clock_reference_time = clock_source->time;
-            }
-            if (status != nullptr) {
-                status->clock_valid = true;
-                status->clock_reference_time = clockReferenceTime(*clock_source);
-                status->ssr_clock_iod = clock_source->ssr_clock_iod;
-            }
-        }
-        if (ura_sigma_m != nullptr && ura_source != nullptr && ura_source->ura_valid) {
-            *ura_sigma_m = ura_source->ura_sigma_m;
-            if (status != nullptr) {
-                status->ura_valid = true;
-                status->ura_reference_time = ura_source->time;
-            }
-        }
-        if (code_bias_m != nullptr && code_source != nullptr && code_source->code_bias_valid) {
-            *code_bias_m = code_source->code_bias_m;
-            if (code_bias_rtklib_m != nullptr) {
-                *code_bias_rtklib_m = code_source->code_bias_rtklib_m;
-            }
-            if (status != nullptr) {
-                status->code_bias_valid = true;
-                status->code_bias_reference_time = code_source->time;
-            }
-        }
-        if (phase_bias_m != nullptr && phase_source != nullptr && phase_source->phase_bias_valid) {
-            *phase_bias_m = phase_source->phase_bias_m;
-            if (phase_bias_rtklib_m != nullptr) {
-                *phase_bias_rtklib_m = phase_source->phase_bias_rtklib_m;
-            }
-            if (phase_bias_discnt != nullptr) {
-                *phase_bias_discnt = phase_source->phase_bias_discnt;
-            }
-            if (phase_bias_reference_time != nullptr) {
-                *phase_bias_reference_time = phase_source->time;
-            }
-            if (status != nullptr) {
-                status->phase_bias_valid = true;
-                status->phase_bias_reference_time = phase_source->time;
-            }
-        }
-        if (atmos_tokens != nullptr && atmos_source != nullptr && atmos_source->atmos_valid) {
-            *atmos_tokens = atmos_source->atmos_tokens;
-            if (atmos_reference_time != nullptr) {
-                *atmos_reference_time = atmos_source->time;
-            }
-            if (status != nullptr) {
-                status->atmos_valid = true;
-                status->atmos_reference_time = atmos_source->time;
-            }
-        }
-        return applyClaslibClockSelection(
-            clock_selection_policy,
-            sat,
-            entries,
-            time,
-            clock_correction_m,
-            base_clock_correction_m,
-            base_clock_valid,
-            clock_reference_time,
-            status,
-            orbit_source != nullptr || clock_source != nullptr || ura_source != nullptr ||
-                code_source != nullptr || phase_source != nullptr || atmos_source != nullptr);
-    }
-
-    const SSROrbitClockCorrection* before = nullptr;
-    const SSROrbitClockCorrection* after = nullptr;
-    if (lower != entries.end()) {
-        after = &(*lower);
-        if (lower != entries.begin()) {
-            before = &(*(lower - 1));
-        }
-    } else {
-        before = &entries.back();
-    }
-
-    if (before == nullptr && after == nullptr) {
-        return false;
-    }
-
-    if (!allow_future_samples && before != nullptr && after != nullptr && before != after) {
-        after = nullptr;
-    }
-
-    if (before != nullptr && after != nullptr && before != after) {
-        const double dt_total = after->time - before->time;
-        const double dt_before = time - before->time;
-        if (std::abs(dt_total) < 1e-9 || std::abs(dt_total) > kPreciseInterpolationGapSeconds) {
-            return false;
-        }
-        const double alpha = dt_before / dt_total;
-        if (before->orbit_valid && after->orbit_valid) {
-            orbit_correction_ecef =
-                before->orbit_correction_ecef +
-                alpha * (after->orbit_correction_ecef - before->orbit_correction_ecef);
-            if (orbit_iode != nullptr) {
-                *orbit_iode = before->iode;
-            }
-            if (status != nullptr) {
-                status->orbit_valid = true;
-                status->orbit_reference_time = orbitReferenceTime(*before);
-                status->orbit_iode = before->iode;
-                status->ssr_orbit_iod = before->ssr_orbit_iod;
-            }
-        } else if (before->orbit_valid) {
-            orbit_correction_ecef = before->orbit_correction_ecef;
-            if (orbit_iode != nullptr) {
-                *orbit_iode = before->iode;
-            }
-            if (status != nullptr) {
-                status->orbit_valid = true;
-                status->orbit_reference_time = orbitReferenceTime(*before);
-                status->orbit_iode = before->iode;
-                status->ssr_orbit_iod = before->ssr_orbit_iod;
-            }
-        } else if (after->orbit_valid) {
-            orbit_correction_ecef = after->orbit_correction_ecef;
-            if (orbit_iode != nullptr) {
-                *orbit_iode = after->iode;
-            }
-            if (status != nullptr) {
-                status->orbit_valid = true;
-                status->orbit_reference_time = orbitReferenceTime(*after);
-                status->orbit_iode = after->iode;
-                status->ssr_orbit_iod = after->ssr_orbit_iod;
-            }
-        } else {
-            orbit_correction_ecef.setZero();
-        }
-        const SSROrbitClockCorrection* held_orbit_source = nullptr;
-        GNSSTime held_orbit_time;
-        auto orbit_end = std::upper_bound(
-            entries.begin(),
-            entries.end(),
-            time,
-            [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                return lhs < rhs.time;
-            });
-        while (orbit_end != entries.begin()) {
-            const SSROrbitClockCorrection& candidate = *--orbit_end;
-            const GNSSTime candidate_time = orbitReferenceTime(candidate);
-            if (candidate_time > time) {
-                continue;
-            }
-            // Compact-SSR reference epochs are causal: an entry cannot refer
-            // to an orbit state newer than its reception/tag epoch. Once the
-            // remaining entry times are older than the best reference epoch,
-            // no earlier candidate can replace it.
-            if (held_orbit_source != nullptr && candidate.time < held_orbit_time) {
-                break;
-            }
-            if (!candidate.orbit_valid) {
-                continue;
-            }
-            if (held_orbit_source == nullptr || candidate_time > held_orbit_time) {
-                held_orbit_source = &candidate;
-                held_orbit_time = candidate_time;
-            }
-        }
-        const bool qzss_orbit_withdrawn =
-            qzssOrbitWithdrawnByNewBank(held_orbit_source);
-        if (qzss_orbit_withdrawn) {
-            orbit_correction_ecef.setZero();
-            if (status != nullptr) {
-                status->orbit_valid = false;
-                status->orbit_withdrawn = true;
-                status->orbit_reference_time = GNSSTime();
-                status->orbit_iode = -1;
-                status->ssr_orbit_iod = -1;
-            }
-        }
-
-        if (before->clock_valid && after->clock_valid) {
-            clock_correction_m =
-                before->clock_correction_m +
-                alpha * (after->clock_correction_m - before->clock_correction_m);
-            if (status != nullptr) {
-                status->clock_valid = true;
-                status->clock_reference_time = clockReferenceTime(*before);
-                status->ssr_clock_iod = before->ssr_clock_iod;
-            }
-        } else if (before->clock_valid) {
-            clock_correction_m = before->clock_correction_m;
-            if (status != nullptr) {
-                status->clock_valid = true;
-                status->clock_reference_time = clockReferenceTime(*before);
-                status->ssr_clock_iod = before->ssr_clock_iod;
-            }
-        } else if (after->clock_valid) {
-            clock_correction_m = after->clock_correction_m;
-            if (status != nullptr) {
-                status->clock_valid = true;
-                status->clock_reference_time = clockReferenceTime(*after);
-                status->ssr_clock_iod = after->ssr_clock_iod;
-            }
-        } else {
-            clock_correction_m = 0.0;
-        }
-        if (before != nullptr || after != nullptr) {
-            interpolateBaseClockOutputs(before, after, alpha, clock_correction_m,
-                                        base_clock_correction_m, base_clock_valid,
-                                        clock_selection_policy);
-        }
-        if (ura_sigma_m != nullptr) {
-            if (before->ura_valid && after->ura_valid) {
-                *ura_sigma_m =
-                    before->ura_sigma_m + alpha * (after->ura_sigma_m - before->ura_sigma_m);
-                if (status != nullptr) {
-                    status->ura_valid = true;
-                    status->ura_reference_time = before->time;
-                }
-            } else if (before->ura_valid) {
-                *ura_sigma_m = before->ura_sigma_m;
-                if (status != nullptr) {
-                    status->ura_valid = true;
-                    status->ura_reference_time = before->time;
-                }
-            } else if (after->ura_valid) {
-                *ura_sigma_m = after->ura_sigma_m;
-                if (status != nullptr) {
-                    status->ura_valid = true;
-                    status->ura_reference_time = after->time;
-                }
-            }
-        }
-        // Bias forward-fill: CLAS broadcasts bias every 30s but clock every 5s.
-        // The interpolation neighbours (before/after) are adjacent SSR entries
-        // which are typically clock-only rows with no bias.  Scan backwards
-        // through older entries to find the most recent bias, matching the
-        // CLASLIB behaviour of holding the last received bias indefinitely.
-        if (code_bias_m != nullptr) {
-            // Code biases are stepwise corrections.  Do not consume a future
-            // bias row just because the next clock interpolation neighbour has
-            // one; CLAS holds the last received code-bias bank until a new
-            // bank is causally available.
-            const SSROrbitClockCorrection* picked = nullptr;
-            if (qzss_atmos_fix) {
-                const size_t scan_end = static_cast<size_t>(lower - entries.begin());
-                picked = findQzssHeldDedicatedBias(false, scan_end);
-            } else if (const auto* scanned = scanBackwardBias(false)) {
-                picked = scanned;
-            } else if (before->code_bias_valid && !before->code_bias_m.empty()) {
-                picked = before;
-            }
-            if (picked != nullptr) {
-                *code_bias_m = picked->code_bias_m;
-                if (code_bias_rtklib_m != nullptr) {
-                    *code_bias_rtklib_m = picked->code_bias_rtklib_m;
-                }
-                if (status != nullptr) {
-                    status->code_bias_valid = true;
-                    status->code_bias_reference_time = picked->time;
-                }
-            }
-        }
-        if (phase_bias_m != nullptr) {
-            const SSROrbitClockCorrection* picked = nullptr;
-            if (!qzss_atmos_fix) {
-                if (clock_selection_policy ==
-                    SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-                    picked = scanBackwardBias(true);
-                } else if (biasScore(*before, true) >= biasScore(*after, true) &&
-                    before->phase_bias_valid && !before->phase_bias_m.empty()) {
-                    picked = before;
-                } else if (after->phase_bias_valid && !after->phase_bias_m.empty()) {
-                    picked = after;
-                } else if (const auto* scanned = scanBackwardBias(true)) {
-                    picked = scanned;
-                }
-            }
-            if (picked != nullptr) {
-                *phase_bias_m = picked->phase_bias_m;
-                if (phase_bias_rtklib_m != nullptr) {
-                    *phase_bias_rtklib_m = picked->phase_bias_rtklib_m;
-                }
-                if (phase_bias_discnt != nullptr) {
-                    *phase_bias_discnt = picked->phase_bias_discnt;
-                }
-                if (phase_bias_reference_time != nullptr) {
-                    *phase_bias_reference_time = picked->time;
-                }
-                if (status != nullptr) {
-                    status->phase_bias_valid = true;
-                    status->phase_bias_reference_time = picked->time;
-                }
-            }
-        }
-        if (clock_reference_time != nullptr) {
-            if (after->clock_valid) {
-                *clock_reference_time = after->time;
-            } else if (before->clock_valid) {
-                *clock_reference_time = before->time;
-            }
-        }
-        if (atmos_tokens != nullptr) {
-            if (qzss_atmos_fix) {
-                const SSROrbitClockCorrection* picked_atmos = pickQzssHeldAtmos();
-                if (picked_atmos != nullptr) {
-                    *atmos_tokens = picked_atmos->atmos_tokens;
-                    if (atmos_reference_time != nullptr) {
-                        *atmos_reference_time = picked_atmos->time;
-                    }
-                    if (status != nullptr) {
-                        status->atmos_valid = true;
-                        status->atmos_reference_time = picked_atmos->time;
-                    }
-                }
-            } else if (atmosScore(*before) >= atmosScore(*after) &&
-                       before->atmos_valid && !before->atmos_tokens.empty()) {
-                *atmos_tokens = before->atmos_tokens;
-                if (atmos_reference_time != nullptr) {
-                    *atmos_reference_time = before->time;
-                }
-                if (status != nullptr) {
-                    status->atmos_valid = true;
-                    status->atmos_reference_time = before->time;
-                }
-            } else if (after->atmos_valid && !after->atmos_tokens.empty()) {
-                *atmos_tokens = after->atmos_tokens;
-                if (atmos_reference_time != nullptr) {
-                    *atmos_reference_time = after->time;
-                }
-                if (status != nullptr) {
-                    status->atmos_valid = true;
-                    status->atmos_reference_time = after->time;
-                }
-            } else {
-                // Neither interpolation neighbour has atmos — scan backwards.
-                auto scan = lower;
-                while (scan != entries.begin()) {
-                    --scan;
-                    if (scan->atmos_valid && !scan->atmos_tokens.empty() &&
-                        std::abs(scan->time - time) <= kPreciseInterpolationGapSeconds) {
-                        *atmos_tokens = scan->atmos_tokens;
-                        if (atmos_reference_time != nullptr) {
-                            *atmos_reference_time = scan->time;
-                        }
-                        if (status != nullptr) {
-                            status->atmos_valid = true;
-                            status->atmos_reference_time = scan->time;
-                        }
-                        break;
-                    }
-                }
-            }
-        }
-        if (qzss_orbit_withdrawn) {
-            return false;
-        }
-        return applyClaslibClockSelection(
-            clock_selection_policy,
-            sat,
-            entries,
-            time,
-            clock_correction_m,
-            base_clock_correction_m,
-            base_clock_valid,
-            clock_reference_time,
-            status,
-            before->orbit_valid || after->orbit_valid ||
-                before->clock_valid || after->clock_valid ||
-                before->ura_valid || after->ura_valid ||
-                before->code_bias_valid || after->code_bias_valid ||
-                before->phase_bias_valid || after->phase_bias_valid ||
-                before->atmos_valid || after->atmos_valid);
-    }
-
-    const SSROrbitClockCorrection* sample = before != nullptr ? before : after;
-    if (sample == nullptr) {
-        return false;
-    }
-    if (std::abs(sample->time - time) > kPreciseInterpolationGapSeconds) {
-        return false;
-    }
-
-    // If the matched sample lacks orbit or clock, scan backwards for a
-    // recent entry that has them.  This handles atmos-only rows that were
-    // inserted at intervening TOWs without orbit/clock data.
-    const SSROrbitClockCorrection* orbit_source = sample;
-    const SSROrbitClockCorrection* clock_source = sample;
-    if (!sample->orbit_valid || !sample->clock_valid) {
-        // Start after the complete sample-time group.  A compact epoch can
-        // contain several service-network/bias variants; lower_bound() starts
-        // at the first one and the old pre-decrement skipped every sibling at
-        // that same TOW.  In literal causal mode this made an orbit row at the
-        // current compact epoch look stale whenever entries.back() for the
-        // group happened to be an atmos/bias-only variant.
-        auto scan = std::upper_bound(
-            entries.begin(), entries.end(), sample->time,
-            [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                return lhs < rhs.time;
-            });
-        while (scan != entries.begin()) {
-            --scan;
-            if (std::abs(scan->time - time) > kPreciseInterpolationGapSeconds) break;
-            if (!orbit_source->orbit_valid && scan->orbit_valid) orbit_source = &(*scan);
-            if (!clock_source->clock_valid && scan->clock_valid) clock_source = &(*scan);
-            if (orbit_source->orbit_valid && clock_source->clock_valid) break;
-        }
-    }
-    const bool orbit_withdrawn = qzssOrbitWithdrawnByNewBank(orbit_source);
-    orbit_correction_ecef =
-        orbit_source->orbit_valid && !orbit_withdrawn
-            ? orbit_source->orbit_correction_ecef
-            : Vector3d::Zero();
-    clock_correction_m = clock_source->clock_valid ? clock_source->clock_correction_m : 0.0;
-    assignBaseClockOutputs(*clock_source, clock_correction_m,
-                           base_clock_correction_m, base_clock_valid);
-    if (orbit_source->orbit_valid && !orbit_withdrawn) {
-        if (orbit_iode != nullptr) {
-            *orbit_iode = orbit_source->iode;
-        }
-        if (status != nullptr) {
-            status->orbit_valid = true;
-            status->orbit_reference_time = orbitReferenceTime(*orbit_source);
-            status->orbit_iode = orbit_source->iode;
-            status->ssr_orbit_iod = orbit_source->ssr_orbit_iod;
-        }
-    }
-    if (orbit_withdrawn && status != nullptr) {
-        status->orbit_withdrawn = true;
-    }
-    if (clock_reference_time != nullptr && clock_source->clock_valid) {
-        *clock_reference_time = clock_source->time;
-    }
-    if (clock_source->clock_valid && status != nullptr) {
-        status->clock_valid = true;
-        status->clock_reference_time = clockReferenceTime(*clock_source);
-        status->ssr_clock_iod = clock_source->ssr_clock_iod;
-    }
-    if (ura_sigma_m != nullptr) {
-        *ura_sigma_m = sample->ura_valid ? sample->ura_sigma_m : 0.0;
-        if (sample->ura_valid && status != nullptr) {
-            status->ura_valid = true;
-            status->ura_reference_time = sample->time;
-        }
-    }
-    if (code_bias_m != nullptr) {
-        const SSROrbitClockCorrection* picked = nullptr;
-        if (qzss_atmos_fix) {
-            const size_t scan_end =
-                lower != entries.end() ? static_cast<size_t>(lower - entries.begin()) + 1
-                                       : entries.size();
-            picked = findQzssHeldDedicatedBias(false, scan_end);
-        } else if (clock_selection_policy == SSRClockSelectionPolicy::ClaslibBaseHold ||
-                   clock_selection_policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-            picked = scanBackwardBias(false);
-        }
-        if (picked == nullptr) {
-            picked = scanBackwardBias(false);
-        }
-        if (picked == nullptr && sample->code_bias_valid && !sample->code_bias_m.empty()) {
-            picked = sample;
-        } else if (picked == nullptr && orbit_source != sample &&
-                   orbit_source->code_bias_valid && !orbit_source->code_bias_m.empty()) {
-            picked = orbit_source;
-        }
-        if (picked != nullptr) {
-            *code_bias_m = picked->code_bias_m;
-            if (code_bias_rtklib_m != nullptr) {
-                *code_bias_rtklib_m = picked->code_bias_rtklib_m;
-            }
-            if (status != nullptr) {
-                status->code_bias_valid = true;
-                status->code_bias_reference_time = picked->time;
-            }
-        }
-    }
-    if (phase_bias_m != nullptr) {
-        const SSROrbitClockCorrection* picked = nullptr;
-        if (!qzss_atmos_fix) {
-            if (clock_selection_policy == SSRClockSelectionPolicy::ClaslibBaseHold ||
-                clock_selection_policy == SSRClockSelectionPolicy::MrtklibLiteralBaseHold) {
-                picked = scanBackwardBias(true);
-            }
-            if (picked == nullptr && sample->phase_bias_valid && !sample->phase_bias_m.empty()) {
-                picked = sample;
-            } else if (picked == nullptr && orbit_source != sample &&
-                       orbit_source->phase_bias_valid && !orbit_source->phase_bias_m.empty()) {
-                picked = orbit_source;
-            }
-        }
-        if (picked != nullptr) {
-            *phase_bias_m = picked->phase_bias_m;
-            if (phase_bias_rtklib_m != nullptr) {
-                *phase_bias_rtklib_m = picked->phase_bias_rtklib_m;
-            }
-            if (phase_bias_discnt != nullptr) {
-                *phase_bias_discnt = picked->phase_bias_discnt;
-            }
-            if (phase_bias_reference_time != nullptr) {
-                *phase_bias_reference_time = picked->time;
-            }
-            if (status != nullptr) {
-                status->phase_bias_valid = true;
-                status->phase_bias_reference_time = picked->time;
-            }
-        }
-    }
-    if (atmos_tokens != nullptr) {
-        if (qzss_atmos_fix) {
-            const SSROrbitClockCorrection* picked_atmos = pickQzssHeldAtmos();
-            if (picked_atmos != nullptr) {
-                *atmos_tokens = picked_atmos->atmos_tokens;
-                if (atmos_reference_time != nullptr) {
-                    *atmos_reference_time = picked_atmos->time;
-                }
-                if (status != nullptr) {
-                    status->atmos_valid = true;
-                    status->atmos_reference_time = picked_atmos->time;
-                }
-            }
-        } else if (sample->atmos_valid) {
-            *atmos_tokens = sample->atmos_tokens;
-            if (atmos_reference_time != nullptr) {
-                *atmos_reference_time = sample->time;
-            }
-            if (status != nullptr) {
-                status->atmos_valid = true;
-                status->atmos_reference_time = sample->time;
-            }
-        } else {
-            // The exact-match sample has no atmos — scan backwards for the
-            // nearest entry that does.  CLAS atmosphere messages arrive at a
-            // lower cadence than orbit/clock, so a recent neighbour is valid.
-            auto it = std::lower_bound(
-                entries.begin(), entries.end(), sample->time,
-                [](const SSROrbitClockCorrection& lhs, const GNSSTime& rhs) {
-                    return lhs.time < rhs;
-                });
-            while (it != entries.begin()) {
-                --it;
-                if (it->atmos_valid && !it->atmos_tokens.empty() &&
-                    std::abs(it->time - time) <= kPreciseInterpolationGapSeconds) {
-                    *atmos_tokens = it->atmos_tokens;
-                    if (atmos_reference_time != nullptr) {
-                        *atmos_reference_time = it->time;
-                    }
-                    if (status != nullptr) {
-                        status->atmos_valid = true;
-                        status->atmos_reference_time = it->time;
-                    }
-                    break;
-                }
-            }
-        }
-    }
-    return applyClaslibClockSelection(
-        clock_selection_policy,
-        sat,
-        entries,
-        time,
-        clock_correction_m,
-        base_clock_correction_m,
-        base_clock_valid,
-        clock_reference_time,
-        status,
-        sample->orbit_valid || sample->clock_valid || sample->ura_valid ||
-            sample->code_bias_valid || sample->phase_bias_valid || sample->atmos_valid);
-}
-
-bool SSRProducts::heldQzssPhaseBiasForServiceNetwork(
-    const SatelliteId& sat,
-    const GNSSTime& time,
-    int service_network_id,
-    std::map<uint8_t, double>* phase_bias_m,
-    std::map<uint8_t, int>* phase_bias_discnt,
-    GNSSTime* phase_bias_reference_time,
-    bool include_equal_time_group,
-    double max_hold_age_seconds) const {
-    if ((!include_equal_time_group &&
-         !pppEnvOverrides().clas_qzss_s_prn_fix) ||
-        sat.system != GNSSSystem::QZSS ||
-        service_network_id <= 0 || phase_bias_m == nullptr) {
-        return false;
-    }
-    return heldClasPhaseBiasForServiceNetwork(
-        sat, time, service_network_id, phase_bias_m, phase_bias_discnt,
-        phase_bias_reference_time, include_equal_time_group,
-        max_hold_age_seconds);
-}
-
-bool SSRProducts::heldClasPhaseBiasForServiceNetwork(
-    const SatelliteId& sat,
-    const GNSSTime& time,
-    int service_network_id,
-    std::map<uint8_t, double>* phase_bias_m,
-    std::map<uint8_t, int>* phase_bias_discnt,
-    GNSSTime* phase_bias_reference_time,
-    bool apply_reception_lag,
-    double max_hold_age_seconds) const {
-    if (service_network_id <= 0 || phase_bias_m == nullptr) {
-        return false;
-    }
-
-    const auto sat_it = orbit_clock_corrections.find(sat);
-    if (sat_it == orbit_clock_corrections.end()) {
-        return false;
-    }
-    const auto& entries = sat_it->second;
-    const GNSSTime selection_time = apply_reception_lag
-        ? time - kMrtklibBiasReceptionLagSeconds
-        : time;
-    auto lower = std::lower_bound(
-        entries.begin(),
-        entries.end(),
-        selection_time,
-        [](const SSROrbitClockCorrection& lhs, const GNSSTime& rhs) {
-            return lhs.time < rhs;
-        });
-    size_t scan_end;
-    if (apply_reception_lag) {
-        const auto causal_end = std::upper_bound(
-            entries.begin(), entries.end(), selection_time,
-            [](const GNSSTime& lhs, const SSROrbitClockCorrection& rhs) {
-                return lhs < rhs.time;
-            });
-        scan_end = static_cast<size_t>(causal_end - entries.begin());
-    } else {
-        scan_end = lower != entries.end()
-            ? static_cast<size_t>(lower - entries.begin()) + 1
-            : entries.size();
-    }
-    const SSROrbitClockCorrection* picked =
-        findQzssHeldPhaseBiasForServiceNetwork(
-            entries, selection_time, service_network_id, scan_end);
-    if (picked == nullptr) {
-        return false;
-    }
-    // A CLAS ST6 network bank is replaced every 30 seconds.  A satellite or
-    // service network omitted from the new bank is an explicit withdrawal,
-    // not permission to retain the previous bank indefinitely.  Keep the
-    // historical unbounded behavior unless the literal caller requests a
-    // bank-age gate.
-    const double hold_age_seconds = selection_time - picked->time;
-    if (max_hold_age_seconds >= 0.0 &&
-        hold_age_seconds >= max_hold_age_seconds - 1e-9) {
-        return false;
-    }
-
-    phase_bias_m->clear();
-    *phase_bias_m = picked->phase_bias_m;
-    if (phase_bias_discnt != nullptr) {
-        *phase_bias_discnt = picked->phase_bias_discnt;
-    }
-    if (phase_bias_reference_time != nullptr) {
-        *phase_bias_reference_time = picked->time;
-    }
-    return !phase_bias_m->empty();
-}
-
-bool SSRProducts::heldAtmosTokensForNetwork(int network_id,
-                                            const GNSSTime& time,
-                                            double max_age_seconds,
-                                            std::map<std::string, std::string>& atmos_tokens,
-                                            GNSSTime* atmos_reference_time) const {
-    if (network_id <= 0) {
-        return false;
-    }
-
-    const SSROrbitClockCorrection* best = nullptr;
-    GNSSTime best_time;
-    for (const auto& sat_entry : orbit_clock_corrections) {
-        const auto first_future = std::upper_bound(
-            sat_entry.second.begin(), sat_entry.second.end(), time,
-            [](const GNSSTime& epoch,
-               const SSROrbitClockCorrection& entry) {
-                return epoch < entry.time;
-            });
-        for (auto entry_it = std::make_reverse_iterator(first_future);
-             entry_it != sat_entry.second.rend(); ++entry_it) {
-            const auto& entry = *entry_it;
-            if (time - entry.time > max_age_seconds + 1e-9) break;
-            if (!entry.atmos_valid || entry.atmos_tokens.empty()) {
-                continue;
-            }
-            if (entry.atmos_network_id != network_id) {
-                continue;
-            }
-            if (best == nullptr || entry.time > best_time) {
-                best = &entry;
-                best_time = entry.time;
-            }
-            break;
-        }
-    }
-    if (best == nullptr) {
-        return false;
-    }
-    atmos_tokens = best->atmos_tokens;
-    if (atmos_reference_time != nullptr) {
-        *atmos_reference_time = best->time;
-    }
-    return true;
-}
-
-bool SSRProducts::heldClasTropTokens(
-    const GNSSTime& time,
-    double max_age_seconds,
-    int network_id,
-    int minimum_grid_count,
-    std::map<std::string, std::string>& atmos_tokens,
-    GNSSTime* atmos_reference_time) const {
-    const SSROrbitClockCorrection* best = nullptr;
-    GNSSTime best_time;
-    for (auto entry_it = clas_trop_bank_corrections.rbegin();
-         entry_it != clas_trop_bank_corrections.rend(); ++entry_it) {
-        const auto& entry = *entry_it;
-        if (entry.time - time > 1e-9) continue;
-        if (time - entry.time > max_age_seconds + 1e-9) break;
-        const int entry_network = parsePositiveIntToken(
-            entry.atmos_tokens, "atmos_trop_network_id");
-        if (network_id > 0 && entry_network != network_id) continue;
-        const auto residual_it =
-            entry.atmos_tokens.find("atmos_trop_residuals_m");
-        if (residual_it == entry.atmos_tokens.end()) continue;
-        const int residual_count = residual_it->second.empty()
-            ? 0
-            : 1 + static_cast<int>(std::count(
-                  residual_it->second.begin(), residual_it->second.end(), ';'));
-        if (minimum_grid_count > 0 && residual_count < minimum_grid_count) continue;
-        best = &entry;
-        best_time = entry.time;
-        break;
-    }
-    for (const auto& sat_entry : orbit_clock_corrections) {
-        // Correction histories are appended in epoch order.  Only the newest
-        // compatible trop row in each satellite history can win, so avoid a
-        // full-history scan on every satellite at every observation epoch.
-        const auto first_future = std::upper_bound(
-            sat_entry.second.begin(), sat_entry.second.end(), time,
-            [](const GNSSTime& epoch,
-               const SSROrbitClockCorrection& entry) {
-                return epoch < entry.time;
-            });
-        for (auto entry_it = std::make_reverse_iterator(first_future);
-             entry_it != sat_entry.second.rend(); ++entry_it) {
-            const auto& entry = *entry_it;
-            if (time - entry.time > max_age_seconds + 1e-9) break;
-            if (!entry.atmos_valid || entry.atmos_tokens.empty()) continue;
-            // atmos_tokens is an ordered map. Jump to the prefix range instead
-            // of walking every string token for every satellite and receiver
-            // epoch; dense CLAS histories make that linear scan dominant.
-            static const std::string trop_prefix = "atmos_trop_";
-            const auto trop_it = entry.atmos_tokens.lower_bound(trop_prefix);
-            const bool has_trop =
-                trop_it != entry.atmos_tokens.end() &&
-                trop_it->first.compare(0, trop_prefix.size(), trop_prefix) == 0;
-            if (!has_trop) continue;
-            if (network_id > 0) {
-                auto network_it =
-                    entry.atmos_tokens.find("atmos_trop_network_id");
-                if (network_it == entry.atmos_tokens.end()) {
-                    network_it = entry.atmos_tokens.find("atmos_network_id");
-                }
-                if (network_it == entry.atmos_tokens.end()) continue;
-                try {
-                    if (std::stoi(network_it->second) != network_id) continue;
-                } catch (const std::exception&) {
-                    continue;
-                }
-            }
-            if (minimum_grid_count > 0) {
-                const auto residual_it =
-                    entry.atmos_tokens.find("atmos_trop_residuals_m");
-                if (residual_it == entry.atmos_tokens.end()) continue;
-                const int residual_count = residual_it->second.empty()
-                    ? 0
-                    : 1 + static_cast<int>(std::count(
-                          residual_it->second.begin(),
-                          residual_it->second.end(), ';'));
-                if (residual_count < minimum_grid_count) continue;
-            }
-            if (best == nullptr || entry.time > best_time) {
-                best = &entry;
-                best_time = entry.time;
-            }
-            break;
-        }
-    }
-    if (best == nullptr) return false;
-    atmos_tokens.clear();
-    const bool bank_only = best->atmos_tokens.find("atmos_trop_bank_only") !=
-                           best->atmos_tokens.end();
-    for (const auto& [key, value] : best->atmos_tokens) {
-        if (!bank_only || key.rfind("atmos_trop_", 0) == 0 ||
-            key == "atmos_network_id" || key == "atmos_grid_count") {
-            atmos_tokens.emplace(key, value);
-        }
-    }
-    // A compact subtype-12 row can carry the surrounding correction row's
-    // network in atmos_network_id while the troposphere payload itself belongs
-    // to atmos_trop_network_id.  This method has already selected the payload
-    // by the latter, and the grid interpolator consumes the former.  Keep the
-    // returned token set self-consistent so a network-7 bank is not evaluated
-    // against (for example) network-11 grid geometry.
-    const int trop_network_id = parsePositiveIntToken(
-        best->atmos_tokens, "atmos_trop_network_id");
-    if (trop_network_id > 0) {
-        atmos_tokens["atmos_network_id"] = std::to_string(trop_network_id);
-    }
-    if (atmos_reference_time != nullptr) {
-        *atmos_reference_time = best->time;
-    }
-    return true;
-}
-
-const std::map<std::string, std::string>* SSRProducts::heldClasAtmosBankTokens(
-    const GNSSTime& time,
-    double max_age_seconds,
-    int network_id,
-    GNSSTime* atmos_reference_time) const {
-    for (auto entry_it = clas_trop_bank_corrections.rbegin();
-         entry_it != clas_trop_bank_corrections.rend(); ++entry_it) {
-        const auto& entry = *entry_it;
-        if (entry.time - time > 1e-9) continue;
-        if (time - entry.time > max_age_seconds + 1e-9) break;
-        if (network_id > 0 && parsePositiveIntToken(
-                entry.atmos_tokens, "atmos_trop_network_id") != network_id) {
-            continue;
-        }
-        if (atmos_reference_time != nullptr) *atmos_reference_time = entry.time;
-        return &entry.atmos_tokens;
-    }
-    return nullptr;
-}
-
-bool SSRProducts::loadCSVFile(const std::string& filename) {
-    std::ifstream input(filename);
-    if (!input.is_open()) {
-        return false;
-    }
-
-    orbit_corrections_are_rac_ = true;
-    bool loaded_any = false;
-    std::string line;
-    while (std::getline(input, line)) {
-        const std::string trimmed = trimCopy(line);
-        if (trimmed.empty() || trimmed[0] == '#') {
-            continue;
-        }
-
-        std::vector<std::string> columns;
-        std::istringstream csv_stream(trimmed);
-        std::string token;
-        while (std::getline(csv_stream, token, ',')) {
-            columns.push_back(trimCopy(token));
-        }
-        if (columns.size() < 7U) {
-            continue;
-        }
-
-        int week = 0;
-        double tow = 0.0;
-        double dx = 0.0;
-        double dy = 0.0;
-        double dz = 0.0;
-        double dclock_m = 0.0;
-        const std::string& sat_token = columns[2];
-        try {
-            week = std::stoi(columns[0]);
-            tow = std::stod(columns[1]);
-            dx = std::stod(columns[3]);
-            dy = std::stod(columns[4]);
-            dz = std::stod(columns[5]);
-            dclock_m = std::stod(columns[6]);
-        } catch (const std::exception&) {
-            continue;
-        }
-
-        SatelliteId satellite;
-        if (!parseCompactSsrSatelliteToken(sat_token, satellite)) {
-            continue;
-        }
-
-        SSROrbitClockCorrection correction;
-        correction.satellite = satellite;
-        correction.time = GNSSTime(week, tow);
-        correction.orbit_correction_ecef = Vector3d(dx, dy, dz);
-        correction.clock_correction_m = dclock_m;
-        // orbit is valid only if at least one component is non-zero
-        correction.orbit_valid =
-            std::isfinite(dx) && std::isfinite(dy) && std::isfinite(dz) &&
-            (std::abs(dx) > 0.0 || std::abs(dy) > 0.0 || std::abs(dz) > 0.0);
-        correction.clock_valid = std::isfinite(dclock_m) && std::abs(dclock_m) > 0.0;
-        correction.clock_withdrawn = !std::isfinite(dclock_m);
-        for (size_t index = 7; index < columns.size(); ++index) {
-            const std::string& extra = columns[index];
-            if (extra.empty()) {
-                continue;
-            }
-            if (extra.rfind("ura_sigma_m=", 0) == 0) {
-                try {
-                    correction.ura_sigma_m = std::stod(extra.substr(12));
-                    correction.ura_valid = std::isfinite(correction.ura_sigma_m);
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("orbit_iode=", 0) == 0) {
-                try {
-                    correction.iode = std::stoi(extra.substr(11));
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("cbias:", 0) == 0) {
-                const auto equal_pos = extra.find('=');
-                if (equal_pos == std::string::npos || equal_pos <= 6) {
-                    continue;
-                }
-                try {
-                    const int signal_id = std::stoi(extra.substr(6, equal_pos - 6));
-                    const double bias_m = std::stod(extra.substr(equal_pos + 1));
-                    if (signal_id >= 0 && signal_id <= 255 && std::isfinite(bias_m)) {
-                        correction.code_bias_m[static_cast<uint8_t>(signal_id)] = bias_m;
-                    }
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("cbias_code:", 0) == 0) {
-                const auto equal_pos = extra.find('=');
-                if (equal_pos == std::string::npos || equal_pos <= 11) {
-                    continue;
-                }
-                try {
-                    const int code = std::stoi(extra.substr(11, equal_pos - 11));
-                    const double bias_m = std::stod(extra.substr(equal_pos + 1));
-                    if (code > 0 && code <= 255) {
-                        correction.code_bias_rtklib_m[static_cast<uint8_t>(code)] = bias_m;
-                    }
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("pbias:", 0) == 0) {
-                const auto equal_pos = extra.find('=');
-                if (equal_pos == std::string::npos || equal_pos <= 6) {
-                    continue;
-                }
-                try {
-                    const int signal_id = std::stoi(extra.substr(6, equal_pos - 6));
-                    const double bias_m = std::stod(extra.substr(equal_pos + 1));
-                    if (signal_id >= 0 && signal_id <= 255 && std::isfinite(bias_m)) {
-                        correction.phase_bias_m[static_cast<uint8_t>(signal_id)] = bias_m;
-                    }
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("pbias_code:", 0) == 0) {
-                const auto equal_pos = extra.find('=');
-                if (equal_pos == std::string::npos || equal_pos <= 11) {
-                    continue;
-                }
-                try {
-                    const int code = std::stoi(extra.substr(11, equal_pos - 11));
-                    const double bias_m = std::stod(extra.substr(equal_pos + 1));
-                    if (code > 0 && code <= 255) {
-                        correction.phase_bias_rtklib_m[static_cast<uint8_t>(code)] = bias_m;
-                    }
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("bias_network_id=", 0) == 0) {
-                try {
-                    correction.bias_network_id = std::max(0, std::stoi(extra.substr(16)));
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("clock_network_id=", 0) == 0) {
-                try {
-                    correction.clock_network_id = std::max(0, std::stoi(extra.substr(17)));
-                } catch (const std::exception&) {
-                }
-                continue;
-            }
-            if (extra.rfind("atmos_", 0) == 0) {
-                const auto equal_pos = extra.find('=');
-                if (equal_pos == std::string::npos || equal_pos <= 6) {
-                    continue;
-                }
-                const std::string key =
-                    remapCompactSsrAtmosKey(extra.substr(0, equal_pos));
-                const std::string value = extra.substr(equal_pos + 1);
-                correction.atmos_tokens[key] = value;
-                if (key == "atmos_network_id") {
-                    try {
-                        correction.atmos_network_id = std::max(0, std::stoi(value));
-                    } catch (const std::exception&) {
-                    }
-                }
-            }
-        }
-        if (correction.atmos_network_id <= 0 && !correction.atmos_tokens.empty()) {
-            correction.atmos_network_id =
-                parsePositiveIntToken(correction.atmos_tokens, "atmos_network_id");
-        }
-        correction.code_bias_valid =
-            !correction.code_bias_m.empty() || !correction.code_bias_rtklib_m.empty();
-        correction.phase_bias_valid =
-            !correction.phase_bias_m.empty() || !correction.phase_bias_rtklib_m.empty();
-        correction.atmos_valid = !correction.atmos_tokens.empty();
-        const auto bank_only_it =
-            correction.atmos_tokens.find("atmos_trop_bank_only");
-        if (bank_only_it != correction.atmos_tokens.end() &&
-            bank_only_it->second == "1") {
-            clas_trop_bank_corrections.push_back(std::move(correction));
-            loaded_any = true;
-            continue;
-        }
-        // Keep atmos-only rows. QZSS CLAS broadcasts network-wide atmosphere
-        // corrections that may arrive without non-zero orbit/clock deltas, and
-        // the PPP loader needs those rows to preserve the atmosphere tokens.
-        addCorrection(correction);
-        loaded_any = true;
-    }
-
-    // Debug: count atmos entries after loading
-    if (pppEnvOverrides().debug) {
-        size_t atmos_entries = 0;
-        size_t total_entries = 0;
-        for (const auto& [sat, entries] : orbit_clock_corrections) {
-            for (const auto& e : entries) {
-                ++total_entries;
-                if (e.atmos_valid) ++atmos_entries;
-            }
-        }
-        std::cerr << "[SSR-LOAD] total_entries=" << total_entries
-                  << " atmos_entries=" << atmos_entries
-                  << " satellites=" << orbit_clock_corrections.size() << "\n";
-    }
-
-    // Forward-fill orbit corrections: CLAS broadcasts orbit every 30s but
-    // clock every 5s.  Carry the last valid orbit forward to clock-only epochs
-    // so all epochs use SSR-corrected satellite positions.
-    for (auto& [sat, entries] : orbit_clock_corrections) {
-        (void)sat;
-        Vector3d last_orbit = Vector3d::Zero();
-        bool has_last_orbit = false;
-        double last_base_clock = 0.0;
-        bool has_last_base_clock = false;
-        GNSSTime last_base_clock_reference_time;
-        for (auto& entry : entries) {
-            if (entry.orbit_valid) {
-                last_orbit = entry.orbit_correction_ecef;
-                has_last_orbit = true;
-            } else if (has_last_orbit && entry.clock_valid) {
-                entry.orbit_correction_ecef = last_orbit;
-                entry.orbit_valid = true;
-            }
-            if (entry.base_clock_valid) {
-                last_base_clock = entry.base_clock_correction_m;
-                has_last_base_clock = true;
-                last_base_clock_reference_time = entry.base_clock_reference_time.week != 0 ?
-                    entry.base_clock_reference_time : entry.time;
-            } else if (has_last_base_clock && entry.clock_valid) {
-                entry.base_clock_correction_m = last_base_clock;
-                entry.base_clock_valid = true;
-                entry.base_clock_reference_time = last_base_clock_reference_time;
-            }
-        }
-    }
-
-    // Forward-fill atmosphere tokens: CLAS broadcasts atmos every 30s.
-    // Carry the last valid atmos forward so all epochs have STEC/trop data.
-    for (auto& [sat, entries] : orbit_clock_corrections) {
-        (void)sat;
-        std::map<std::string, std::string> last_atmos;
-        int last_atmos_network_id = 0;
-        bool has_last_atmos = false;
-        for (auto& entry : entries) {
-            if (entry.atmos_valid && !entry.atmos_tokens.empty()) {
-                last_atmos = entry.atmos_tokens;
-                last_atmos_network_id = entry.atmos_network_id;
-                has_last_atmos = true;
-            } else if (has_last_atmos && !entry.atmos_valid && entry.clock_valid &&
-                       last_atmos.find("atmos_lifecycle") == last_atmos.end()) {
-                entry.atmos_tokens = last_atmos;
-                entry.atmos_network_id = last_atmos_network_id;
-                entry.atmos_valid = true;
-            }
-        }
-    }
-
-    return loaded_any;
-}
-
-bool SSRProducts::hasData(const SatelliteId& sat, const GNSSTime& time) const {
-    Vector3d orbit_correction = Vector3d::Zero();
-    double clock_correction_m = 0.0;
-    return interpolateCorrection(sat, time, orbit_correction, clock_correction_m);
-}
-
-void SSRProducts::clear() {
-    orbit_clock_corrections.clear();
-    clas_trop_bank_corrections.clear();
-    orbit_corrections_are_rac_ = false;
-}
-
-bool IONEXProducts::loadIONEXFile(const std::string& filename) {
-    clear();
-
-    std::ifstream input(filename);
-    if (!input.is_open()) {
-        return false;
-    }
-
-    bool in_header = true;
-    bool loaded_any = false;
-    bool current_is_rms = false;
-    IONEXMap current_map;
-    std::string line;
-    while (std::getline(input, line)) {
-        if (in_header) {
-            if (line.find("IONEX VERSION / TYPE") != std::string::npos) {
-                const std::string version_text = trimCopy(line.substr(0, std::min<size_t>(20U, line.size())));
-                if (!version_text.empty()) {
-                    std::istringstream version_stream(version_text);
-                    version_stream >> version;
-                }
-                const auto type_fields = trimCopy(line.substr(20, std::min<size_t>(40U, line.size() > 20 ? line.size() - 20U : 0U)));
-                if (!type_fields.empty()) {
-                    std::istringstream type_stream(type_fields);
-                    std::string unused_type;
-                    type_stream >> unused_type >> system;
-                }
-            } else if (line.find("INTERVAL") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 20U, values) && !values.empty()) {
-                    interval_s = static_cast<int>(std::llround(values.front()));
-                }
-            } else if (line.find("MAP DIMENSION") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 20U, values) && !values.empty()) {
-                    map_dimension = static_cast<int>(std::llround(values.front()));
-                }
-            } else if (line.find("BASE RADIUS") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 20U, values) && !values.empty()) {
-                    base_radius_km = values.front();
-                }
-            } else if (line.find("ELEVATION CUTOFF") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 20U, values) && !values.empty()) {
-                    elevation_cutoff_deg = values.front();
-                }
-            } else if (line.find("MAPPING FUNCTION") != std::string::npos) {
-                mapping_function = trimCopy(line.substr(0, std::min<size_t>(20U, line.size())));
-            } else if (line.find("EXPONENT") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 20U, values) && !values.empty()) {
-                    exponent = static_cast<int>(std::llround(values.front()));
-                }
-            } else if (line.find("LAT1 / LAT2 / DLAT") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 30U, values) && values.size() >= 3U) {
-                    latitude_grid = {values[0], values[1], values[2]};
-                }
-            } else if (line.find("LON1 / LON2 / DLON") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 30U, values) && values.size() >= 3U) {
-                    longitude_grid = {values[0], values[1], values[2]};
-                }
-            } else if (line.find("HGT1 / HGT2 / DHGT") != std::string::npos) {
-                std::vector<double> values;
-                if (parseFixedFieldDoubles(line, 30U, values) && values.size() >= 3U) {
-                    height_grid = {values[0], values[1], values[2]};
-                }
-            } else if (line.find("PRN / BIAS / RMS") != std::string::npos) {
-                ++auxiliary_dcb_entries;
-            } else if (line.find("END OF HEADER") != std::string::npos) {
-                in_header = false;
-            }
-            continue;
-        }
-
-        if (line.find("START OF TEC MAP") != std::string::npos) {
-            current_map = IONEXMap{};
-            current_is_rms = false;
-            continue;
-        }
-        if (line.find("START OF RMS MAP") != std::string::npos) {
-            current_map = IONEXMap{};
-            current_is_rms = true;
-            continue;
-        }
-        if (line.find("END OF TEC MAP") != std::string::npos ||
-            line.find("END OF RMS MAP") != std::string::npos) {
-            if (!current_map.rows.empty()) {
-                auto& target = current_is_rms ? rms_maps : tec_maps;
-                target.push_back(current_map);
-                loaded_any = true;
-            }
-            current_map = IONEXMap{};
-            current_is_rms = false;
-            continue;
-        }
-        if (line.find("EPOCH OF CURRENT MAP") != std::string::npos) {
-            parseIonexEpochLine(line, current_map.time);
-            continue;
-        }
-        if (line.find("LAT/LON1/LON2/DLON/H") == std::string::npos) {
-            continue;
-        }
-
-        std::vector<double> row_header;
-        if (!parseFixedFieldDoubles(line, 40U, row_header) || row_header.size() < 5U) {
-            continue;
-        }
-
-        IONEXLatitudeRow row;
-        row.latitude_deg = row_header[0];
-        row.longitude_start_deg = row_header[1];
-        row.longitude_end_deg = row_header[2];
-        row.longitude_step_deg = row_header[3];
-        row.height_km = row_header[4];
-
-        if (std::abs(row.longitude_step_deg) < 1e-12) {
-            continue;
-        }
-
-        const int longitude_count = static_cast<int>(
-            std::llround((row.longitude_end_deg - row.longitude_start_deg) / row.longitude_step_deg)) + 1;
-        if (longitude_count <= 0) {
-            continue;
-        }
-
-        while (static_cast<int>(row.values_tecu.size()) < longitude_count && std::getline(input, line)) {
-            if (line.find("END OF TEC MAP") != std::string::npos ||
-                line.find("END OF RMS MAP") != std::string::npos ||
-                line.find("LAT/LON1/LON2/DLON/H") != std::string::npos) {
-                break;
-            }
-            // IONEX data values are 5-char wide (I5 in the spec) and
-            // packed up to 16 values per 80-character line. The
-            // previous 60-char window dropped the last 4 values on
-            // every line, leaving every TEC row 16-values short and
-            // every map empty — silently rejecting CODE / IGS final
-            // IONEX. Use the full line.
-            std::istringstream value_stream(line);
-            double raw_value = 0.0;
-            while (value_stream >> raw_value) {
-                if (std::abs(raw_value - 9999.0) < 1e-6) {
-                    row.values_tecu.push_back(std::numeric_limits<double>::quiet_NaN());
-                } else {
-                    row.values_tecu.push_back(raw_value * std::pow(10.0, exponent));
-                }
-                if (static_cast<int>(row.values_tecu.size()) >= longitude_count) {
-                    break;
-                }
-            }
-        }
-        if (static_cast<int>(row.values_tecu.size()) == longitude_count) {
-            current_map.rows.push_back(std::move(row));
-        }
-    }
-
-    auto sort_by_time = [](std::vector<IONEXMap>& maps) {
-        std::sort(
-            maps.begin(),
-            maps.end(),
-            [](const IONEXMap& lhs, const IONEXMap& rhs) {
-                return lhs.time < rhs.time;
-            });
-    };
-    sort_by_time(tec_maps);
-    sort_by_time(rms_maps);
-    return loaded_any;
-}
-
-bool IONEXProducts::interpolateTecu(const GNSSTime& time,
-                                    double latitude_deg,
-                                    double longitude_deg,
-                                    double& tecu,
-                                    double* rms_tecu) const {
-    const auto interpolate_map_series =
-        [&](const std::vector<IONEXMap>& maps, double& output_value) -> bool {
-            if (maps.empty()) {
-                return false;
-            }
-
-            const IONEXMap* before = findEntryAtOrBefore(maps, time);
-            const IONEXMap* after = findEntryAtOrAfter(maps, time);
-            if (before == nullptr && after == nullptr) {
-                return false;
-            }
-            if (before != nullptr && after != nullptr && before != after) {
-                double left = 0.0;
-                double right = 0.0;
-                if (!interpolateIonexMap(*before, latitude_deg, longitude_deg, left) ||
-                    !interpolateIonexMap(*after, latitude_deg, longitude_deg, right)) {
-                    return false;
-                }
-                const double dt_total = after->time - before->time;
-                if (std::abs(dt_total) < 1e-9) {
-                    output_value = left;
-                    return true;
-                }
-                const double alpha = (time - before->time) / dt_total;
-                output_value = left + alpha * (right - left);
-                return true;
-            }
-
-            const IONEXMap* sample = before != nullptr ? before : after;
-            return sample != nullptr && interpolateIonexMap(*sample, latitude_deg, longitude_deg, output_value);
-        };
-
-    if (!interpolate_map_series(tec_maps, tecu)) {
-        return false;
-    }
-    if (rms_tecu != nullptr) {
-        double rms_value = 0.0;
-        if (interpolate_map_series(rms_maps, rms_value)) {
-            *rms_tecu = rms_value;
-        } else {
-            *rms_tecu = 0.0;
-        }
-    }
-    return true;
-}
-
-bool IONEXProducts::hasData(const GNSSTime& time) const {
-    return findEntryAtOrBefore(tec_maps, time) != nullptr ||
-           findEntryAtOrAfter(tec_maps, time) != nullptr;
-}
-
-void IONEXProducts::clear() {
-    version.clear();
-    system.clear();
-    interval_s = 0;
-    map_dimension = 0;
-    exponent = 0;
-    base_radius_km = 0.0;
-    elevation_cutoff_deg = 0.0;
-    mapping_function.clear();
-    latitude_grid.clear();
-    longitude_grid.clear();
-    height_grid.clear();
-    auxiliary_dcb_entries = 0;
-    tec_maps.clear();
-    rms_maps.clear();
-}
-
-bool DCBProducts::loadFile(const std::string& filename) {
-    clear();
-
-    std::ifstream input(filename);
-    if (!input.is_open()) {
-        return false;
-    }
-
-    bool in_bias_solution = false;
-    bool loaded_any = false;
-    std::string line;
-    while (std::getline(input, line)) {
-        const std::string trimmed = trimCopy(line);
-        if (trimmed.empty()) {
-            continue;
-        }
-        if (trimmed.rfind("+BIAS/SOLUTION", 0) == 0) {
-            in_bias_solution = true;
-            continue;
-        }
-        if (trimmed.rfind("-BIAS/SOLUTION", 0) == 0) {
-            in_bias_solution = false;
-            continue;
-        }
-
-        if (in_bias_solution) {
-            if (trimmed[0] == '*') {
-                continue;
-            }
-            std::vector<std::string> fields;
-            std::istringstream stream(trimmed);
-            std::string token;
-            while (stream >> token) {
-                fields.push_back(token);
-            }
-            if (fields.size() < 9U) {
-                continue;
-            }
-
-            // Bias-SINEX rows are `BIAS SVN PRN [STATION] OBS1 OBS2
-            // START END UNIT VALUE STDEV`. Some emitters omit the
-            // optional STATION column entirely; some include it as
-            // a 9-char field that whitespace-tokenization collapses
-            // away. After collapsing, the PRN column may therefore
-            // land at fields[1] (no SVN, e.g. CAS) or fields[2]
-            // (BSX with SVN, e.g. GFZ/GBM). Pick whichever parses
-            // as a valid PRN (3-char `<sys><digit><digit>`); SVNs
-            // are 4 chars (`<sys><digit><digit><digit>`) so
-            // length disambiguates without needing brittle field-
-            // count heuristics.
-            SatelliteId satellite;
-            std::size_t obs_offset = 0;
-            if (fields[1].size() == 3U &&
-                parseSatelliteToken(fields[1], satellite)) {
-                obs_offset = 2;
-            } else if (fields[2].size() == 3U &&
-                       parseSatelliteToken(fields[2], satellite)) {
-                obs_offset = 3;
-            } else {
-                continue;
-            }
-            // Required columns (relative to obs_offset): OBS1,
-            // OBS2, START, END, UNIT, VALUE, STDEV — 7 more.
-            if (fields.size() < obs_offset + 7U) {
-                continue;
-            }
-
-            try {
-                DCBEntry entry;
-                entry.bias_type = fields[0];
-                entry.satellite = satellite;
-                entry.observation_1 = fields[obs_offset];
-                entry.observation_2 = fields[obs_offset + 1];
-                entry.unit = fields[obs_offset + 4];
-                entry.bias = std::stod(fields[obs_offset + 5]);
-                entry.sigma = std::stod(fields[obs_offset + 6]);
-                entry.valid = std::isfinite(entry.bias);
-                entries.push_back(entry);
-                loaded_any = true;
-            } catch (const std::exception&) {
-            }
-            continue;
-        }
-
-        if (line.find("PRN / BIAS / RMS") != std::string::npos) {
-            std::vector<double> values;
-            std::vector<std::string> fields;
-            std::istringstream stream(line.substr(0, std::min<size_t>(40U, line.size())));
-            std::string token;
-            while (stream >> token) {
-                fields.push_back(token);
-            }
-            if (fields.size() < 3U) {
-                continue;
-            }
-            SatelliteId satellite;
-            if (!parseSatelliteToken(fields[0], satellite)) {
-                continue;
-            }
-            try {
-                DCBEntry entry;
-                entry.bias_type = "DCB";
-                entry.satellite = satellite;
-                entry.unit = "ns";
-                entry.bias = std::stod(fields[1]);
-                entry.sigma = std::stod(fields[2]);
-                entry.valid = std::isfinite(entry.bias);
-                entries.push_back(entry);
-                loaded_any = true;
-            } catch (const std::exception&) {
-            }
-        }
-    }
-
-    return loaded_any;
-}
-
-bool DCBProducts::getBias(const SatelliteId& sat,
-                          const std::string& bias_type,
-                          const std::string& observation_1,
-                          const std::string& observation_2,
-                          double& bias,
-                          double* sigma) const {
-    for (const auto& entry : entries) {
-        if (!entry.valid || !(entry.satellite == sat) || entry.bias_type != bias_type) {
-            continue;
-        }
-        if (!observation_1.empty() && entry.observation_1 != observation_1) {
-            continue;
-        }
-        if (!observation_2.empty() && entry.observation_2 != observation_2) {
-            continue;
-        }
-        bias = entry.bias;
-        if (sigma != nullptr) {
-            *sigma = entry.sigma;
-        }
-        return true;
-    }
-    return false;
-}
-
-void DCBProducts::clear() {
-    entries.clear();
-}
-
 } // namespace libgnss
